Module elasticsearch::autoscaling

Autoscaling APIs

The autoscaling feature enables an operator to configure tiers of nodes that self-monitor whether or not they need to scale based on an operator-defined policy.

Structs

Autoscaling

Namespace client for Autoscaling APIs

AutoscalingDeleteAutoscalingPolicy

Builder for the Autoscaling Delete Autoscaling Policy API

AutoscalingGetAutoscalingCapacity

Builder for the Autoscaling Get Autoscaling Capacity API

AutoscalingGetAutoscalingPolicy

Builder for the Autoscaling Get Autoscaling Policy API

AutoscalingPutAutoscalingPolicy

Builder for the Autoscaling Put Autoscaling Policy API

Enums

AutoscalingDeleteAutoscalingPolicyParts

API parts for the Autoscaling Delete Autoscaling Policy API

AutoscalingGetAutoscalingCapacityParts

API parts for the Autoscaling Get Autoscaling Capacity API

AutoscalingGetAutoscalingPolicyParts

API parts for the Autoscaling Get Autoscaling Policy API

AutoscalingPutAutoscalingPolicyParts

API parts for the Autoscaling Put Autoscaling Policy API