Module elasticsearch::ccr

Cross-cluster Replication APIs

Enable replication of indices in remote clusters to a local cluster. This functionality can be used in some common production use cases:

Disaster recovery in case a primary cluster fails. A secondary cluster can serve as a hot backup

Geo-proximity so that reads can be served locally

Structs

Ccr

Namespace client for Cross Cluster Replication APIs

CcrDeleteAutoFollowPattern

Builder for the Ccr Delete Auto Follow Pattern API

CcrFollow

Builder for the Ccr Follow API

CcrFollowInfo

Builder for the Ccr Follow Info API

CcrFollowStats

Builder for the Ccr Follow Stats API

CcrForgetFollower

Builder for the Ccr Forget Follower API

CcrGetAutoFollowPattern

Builder for the Ccr Get Auto Follow Pattern API

CcrPauseAutoFollowPattern

Builder for the Ccr Pause Auto Follow Pattern API

CcrPauseFollow

Builder for the Ccr Pause Follow API

CcrPutAutoFollowPattern

Builder for the Ccr Put Auto Follow Pattern API

CcrResumeAutoFollowPattern

Builder for the Ccr Resume Auto Follow Pattern API

CcrResumeFollow

Builder for the Ccr Resume Follow API

CcrStats

Builder for the Ccr Stats API

CcrUnfollow

Builder for the Ccr Unfollow API

Enums

CcrDeleteAutoFollowPatternParts

API parts for the Ccr Delete Auto Follow Pattern API

CcrFollowInfoParts

API parts for the Ccr Follow Info API

CcrFollowParts

API parts for the Ccr Follow API

CcrFollowStatsParts

API parts for the Ccr Follow Stats API

CcrForgetFollowerParts

API parts for the Ccr Forget Follower API

CcrGetAutoFollowPatternParts

API parts for the Ccr Get Auto Follow Pattern API

CcrPauseAutoFollowPatternParts

API parts for the Ccr Pause Auto Follow Pattern API

CcrPauseFollowParts

API parts for the Ccr Pause Follow API

CcrPutAutoFollowPatternParts

API parts for the Ccr Put Auto Follow Pattern API

CcrResumeAutoFollowPatternParts

API parts for the Ccr Resume Auto Follow Pattern API

CcrResumeFollowParts

API parts for the Ccr Resume Follow API

CcrStatsParts

API parts for the Ccr Stats API

CcrUnfollowParts

API parts for the Ccr Unfollow API