Struct elasticsearch::ccr::Ccr

pub struct Ccr<'a> { /* private fields */ }

Namespace client for Cross Cluster Replication APIs

Implementations

impl<'a> Ccr<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Ccr

pub fn transport(&self) -> &Transport

pub fn delete_auto_follow_pattern<'b>(
&'a self,
parts: CcrDeleteAutoFollowPatternParts<'b>
) -> CcrDeleteAutoFollowPattern<'a, 'b>

Ccr Delete Auto Follow Pattern API

Deletes auto-follow patterns.

pub fn follow<'b>(&'a self, parts: CcrFollowParts<'b>) -> CcrFollow<'a, 'b, ()>

Ccr Follow API

Creates a new follower index configured to follow the referenced leader index.

pub fn follow_info<'b>(
&'a self,
parts: CcrFollowInfoParts<'b>
) -> CcrFollowInfo<'a, 'b>

Ccr Follow Info API

Retrieves information about all follower indices, including parameters and status for each follower index

pub fn follow_stats<'b>(
&'a self,
parts: CcrFollowStatsParts<'b>
) -> CcrFollowStats<'a, 'b>

Ccr Follow Stats API

Retrieves follower stats. return shard-level stats about the following tasks associated with each shard for the specified indices.

pub fn forget_follower<'b>(
&'a self,
parts: CcrForgetFollowerParts<'b>
) -> CcrForgetFollower<'a, 'b, ()>

Ccr Forget Follower API

Removes the follower retention leases from the leader.

pub fn get_auto_follow_pattern<'b>(
&'a self,
parts: CcrGetAutoFollowPatternParts<'b>
) -> CcrGetAutoFollowPattern<'a, 'b>

Ccr Get Auto Follow Pattern API

Gets configured auto-follow patterns. Returns the specified auto-follow pattern collection.

pub fn pause_auto_follow_pattern<'b>(
&'a self,
parts: CcrPauseAutoFollowPatternParts<'b>
) -> CcrPauseAutoFollowPattern<'a, 'b, ()>

Ccr Pause Auto Follow Pattern API

Pauses an auto-follow pattern

pub fn pause_follow<'b>(
&'a self,
parts: CcrPauseFollowParts<'b>
) -> CcrPauseFollow<'a, 'b, ()>

Ccr Pause Follow API

Pauses a follower index. The follower index will not fetch any additional operations from the leader index.

pub fn put_auto_follow_pattern<'b>(
&'a self,
parts: CcrPutAutoFollowPatternParts<'b>
) -> CcrPutAutoFollowPattern<'a, 'b, ()>

Ccr Put Auto Follow Pattern API

Creates a new named collection of auto-follow patterns against a specified remote cluster. Newly created indices on the remote cluster matching any of the specified patterns will be automatically configured as follower indices.

pub fn resume_auto_follow_pattern<'b>(
&'a self,
parts: CcrResumeAutoFollowPatternParts<'b>
) -> CcrResumeAutoFollowPattern<'a, 'b, ()>

Ccr Resume Auto Follow Pattern API

Resumes an auto-follow pattern that has been paused

pub fn resume_follow<'b>(
&'a self,
parts: CcrResumeFollowParts<'b>
) -> CcrResumeFollow<'a, 'b, ()>

Ccr Resume Follow API

Resumes a follower index that has been paused

pub fn stats<'b>(&'a self) -> CcrStats<'a, 'b>

Ccr Stats API

Gets all stats related to cross-cluster replication.

pub fn unfollow<'b>(
&'a self,
parts: CcrUnfollowParts<'b>
) -> CcrUnfollow<'a, 'b, ()>

Ccr Unfollow API

Stops the following task associated with a follower index and removes index metadata and settings associated with cross-cluster replication.