Module elasticsearch::cluster

Cluster APIs

Manage settings, perform operations, and retrieve information about an Elasticsearch cluster.

Structs

Cluster

Namespace client for Cluster APIs

ClusterAllocationExplain

Builder for the Cluster Allocation Explain API

ClusterDeleteComponentTemplate

Builder for the Cluster Delete Component Template API

ClusterDeleteVotingConfigExclusions

Builder for the Cluster Delete Voting Config Exclusions API

ClusterExistsComponentTemplate

Builder for the Cluster Exists Component Template API

ClusterGetComponentTemplate

Builder for the Cluster Get Component Template API

ClusterGetSettings

Builder for the Cluster Get Settings API

ClusterHealth

Builder for the Cluster Health API

ClusterPendingTasks

Builder for the Cluster Pending Tasks API

ClusterPostVotingConfigExclusions

Builder for the Cluster Post Voting Config Exclusions API

ClusterPutComponentTemplate

Builder for the Cluster Put Component Template API

ClusterPutSettings

Builder for the Cluster Put Settings API

ClusterRemoteInfo

Builder for the Cluster Remote Info API

ClusterReroute

Builder for the Cluster Reroute API

ClusterState

Builder for the Cluster State API

ClusterStats

Builder for the Cluster Stats API

Enums

ClusterAllocationExplainParts

API parts for the Cluster Allocation Explain API

ClusterDeleteComponentTemplateParts

API parts for the Cluster Delete Component Template API

ClusterDeleteVotingConfigExclusionsParts

API parts for the Cluster Delete Voting Config Exclusions API

ClusterExistsComponentTemplateParts

API parts for the Cluster Exists Component Template API

ClusterGetComponentTemplateParts

API parts for the Cluster Get Component Template API

ClusterGetSettingsParts

API parts for the Cluster Get Settings API

ClusterHealthParts

API parts for the Cluster Health API

ClusterPendingTasksParts

API parts for the Cluster Pending Tasks API

ClusterPostVotingConfigExclusionsParts

API parts for the Cluster Post Voting Config Exclusions API

ClusterPutComponentTemplateParts

API parts for the Cluster Put Component Template API

ClusterPutSettingsParts

API parts for the Cluster Put Settings API

ClusterRemoteInfoParts

API parts for the Cluster Remote Info API

ClusterRerouteParts

API parts for the Cluster Reroute API

ClusterStateParts

API parts for the Cluster State API

ClusterStatsParts

API parts for the Cluster Stats API