Struct elasticsearch::cluster::Cluster

pub struct Cluster<'a> { /* private fields */ }

Namespace client for Cluster APIs

Implementations

impl<'a> Cluster<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Cluster

pub fn transport(&self) -> &Transport

pub fn allocation_explain<'b>(&'a self) -> ClusterAllocationExplain<'a, 'b, ()>

Cluster Allocation Explain API

Provides explanations for shard allocations in the cluster.

pub fn delete_component_template<'b>(
&'a self,
parts: ClusterDeleteComponentTemplateParts<'b>
) -> ClusterDeleteComponentTemplate<'a, 'b>

Cluster Delete Component Template API

Deletes a component template

pub fn delete_voting_config_exclusions<'b>(
&'a self
) -> ClusterDeleteVotingConfigExclusions<'a, 'b>

Cluster Delete Voting Config Exclusions API

Clears cluster voting config exclusions.

pub fn exists_component_template<'b>(
&'a self,
parts: ClusterExistsComponentTemplateParts<'b>
) -> ClusterExistsComponentTemplate<'a, 'b>

Cluster Exists Component Template API

Returns information about whether a particular component template exist

pub fn get_component_template<'b>(
&'a self,
parts: ClusterGetComponentTemplateParts<'b>
) -> ClusterGetComponentTemplate<'a, 'b>

Cluster Get Component Template API

Returns one or more component templates

pub fn get_settings<'b>(&'a self) -> ClusterGetSettings<'a, 'b>

Cluster Get Settings API

Returns cluster settings.

pub fn health<'b>(
&'a self,
parts: ClusterHealthParts<'b>
) -> ClusterHealth<'a, 'b>

Cluster Health API

Returns basic information about the health of the cluster.

pub fn pending_tasks<'b>(&'a self) -> ClusterPendingTasks<'a, 'b>

Cluster Pending Tasks API

Returns a list of any cluster-level changes (e.g. create index, update mapping, allocate or fail shard) which have not yet been executed.

pub fn post_voting_config_exclusions<'b>(
&'a self
) -> ClusterPostVotingConfigExclusions<'a, 'b, ()>

Cluster Post Voting Config Exclusions API

Updates the cluster voting config exclusions by node ids or node names.

pub fn put_component_template<'b>(
&'a self,
parts: ClusterPutComponentTemplateParts<'b>
) -> ClusterPutComponentTemplate<'a, 'b, ()>

Cluster Put Component Template API

Creates or updates a component template

pub fn put_settings<'b>(&'a self) -> ClusterPutSettings<'a, 'b, ()>

Cluster Put Settings API

Updates the cluster settings.

pub fn remote_info<'b>(&'a self) -> ClusterRemoteInfo<'a, 'b>

Cluster Remote Info API

Returns the information about configured remote clusters.

pub fn reroute<'b>(&'a self) -> ClusterReroute<'a, 'b, ()>

Cluster Reroute API

Allows to manually change the allocation of individual shards in the cluster.

pub fn state<'b>(&'a self, parts: ClusterStateParts<'b>) -> ClusterState<'a, 'b>

Cluster State API

Returns a comprehensive information about the state of the cluster.

pub fn stats<'b>(&'a self, parts: ClusterStatsParts<'b>) -> ClusterStats<'a, 'b>

Cluster Stats API

Returns high-level overview of cluster statistics.