Module elasticsearch::enrich

Enrich APIs

Manage enrich policies that can be used by the enrich processor as part of an ingest pipeline, to add data from your existing indices to incoming documents during ingest.

Structs

Enrich

Namespace client for Enrich APIs

EnrichDeletePolicy

Builder for the Enrich Delete Policy API

EnrichExecutePolicy

Builder for the Enrich Execute Policy API

EnrichGetPolicy

Builder for the Enrich Get Policy API

EnrichPutPolicy

Builder for the Enrich Put Policy API

EnrichStats

Builder for the Enrich Stats API

Enums

EnrichDeletePolicyParts

API parts for the Enrich Delete Policy API

EnrichExecutePolicyParts

API parts for the Enrich Execute Policy API

EnrichGetPolicyParts

API parts for the Enrich Get Policy API

EnrichPutPolicyParts

API parts for the Enrich Put Policy API

EnrichStatsParts

API parts for the Enrich Stats API