Struct elasticsearch::http::response::Error

pub struct Error { /* fields omitted */ }

Details about the exception raised by Elasticsearch

Implementations

impl Error

pub fn caused_by(&self) -> Option<&Cause>

The cause of the exception

pub fn root_cause(&self) -> &Vec<Cause>

The root causes for the exception

pub fn header(&self) -> &BTreeMap<String, Vec<String>>

The headers for the exception

pub fn reason(&self) -> Option<&str>

The reason for the exception, if available.

pub fn stack_trace(&self) -> Option<&str>

The exception stack trace, if available.

Available if error_trace is specified on the request

pub fn ty(&self) -> Option<&str>

The type of exception, if available.

pub fn additional_details(&self) -> &BTreeMap<String, Value>

Additional details about the cause.

Elasticsearch can return additional details about an exception, depending on context, which do not map to fields on Error. These are collected here