Trait elasticsearch::http::transport::ConnectionPool

pub trait ConnectionPool: Debug + DynClone + Sync + Send {
    fn next(&self) -> &Connection;
}

A pool of Connections, used to make API calls to Elasticsearch.

A ConnectionPool manages the connections, with different implementations determining how to get the next Connection. The simplest type of ConnectionPool is SingleNodeConnectionPool, which manages only a single connection, but other implementations may manage connections more dynamically at runtime, based upon the response to API calls.

Required methods

fn next(&self) -> &Connection

Gets a reference to the next Connection