Module elasticsearch::ilm

Index Lifecycle Management APIs

Automate how indices are managed over time. Rather than simply performing management actions on indices on a set schedule, actions can be based on other factors such as shard size and performance requirements.

Control how indices are handled as they age by attaching a lifecycle policy to the index template used to create them. Update the policy to modify the lifecycle of both new and existing indices.

Structs

Ilm

Namespace client for Index Lifecycle Management APIs

IlmDeleteLifecycle

Builder for the Ilm Delete Lifecycle API

IlmExplainLifecycle

Builder for the Ilm Explain Lifecycle API

IlmGetLifecycle

Builder for the Ilm Get Lifecycle API

IlmGetStatus

Builder for the Ilm Get Status API

IlmMigrateToDataTiers

Builder for the Ilm Migrate To Data Tiers API

IlmMoveToStep

Builder for the Ilm Move To Step API

IlmPutLifecycle

Builder for the Ilm Put Lifecycle API

IlmRemovePolicy

Builder for the Ilm Remove Policy API

IlmRetry

Builder for the Ilm Retry API

IlmStart

Builder for the Ilm Start API

IlmStop

Builder for the Ilm Stop API

Enums

IlmDeleteLifecycleParts

API parts for the Ilm Delete Lifecycle API

IlmExplainLifecycleParts

API parts for the Ilm Explain Lifecycle API

IlmGetLifecycleParts

API parts for the Ilm Get Lifecycle API

IlmGetStatusParts

API parts for the Ilm Get Status API

IlmMigrateToDataTiersParts

API parts for the Ilm Migrate To Data Tiers API

IlmMoveToStepParts

API parts for the Ilm Move To Step API

IlmPutLifecycleParts

API parts for the Ilm Put Lifecycle API

IlmRemovePolicyParts

API parts for the Ilm Remove Policy API

IlmRetryParts

API parts for the Ilm Retry API

IlmStartParts

API parts for the Ilm Start API

IlmStopParts

API parts for the Ilm Stop API