Module elasticsearch::indices

Index APIs

Manage individual indices, index settings, aliases, mappings, and index templates.

Structs

Indices

Namespace client for Indices APIs

IndicesAddBlock

Builder for the Indices Add Block API

IndicesAnalyze

Builder for the Indices Analyze API

IndicesClearCache

Builder for the Indices Clear Cache API

IndicesClone

Builder for the Indices Clone API

IndicesClose

Builder for the Indices Close API

IndicesCreate

Builder for the Indices Create API

IndicesCreateDataStream

Builder for the Indices Create Data Stream API

IndicesDataStreamsStats

Builder for the Indices Data Streams Stats API

IndicesDelete

Builder for the Indices Delete API

IndicesDeleteAlias

Builder for the Indices Delete Alias API

IndicesDeleteDataStream

Builder for the Indices Delete Data Stream API

IndicesDeleteIndexTemplate

Builder for the Indices Delete Index Template API

IndicesDeleteTemplate

Builder for the Indices Delete Template API

IndicesExists

Builder for the Indices Exists API

IndicesExistsAlias

Builder for the Indices Exists Alias API

IndicesExistsIndexTemplate

Builder for the Indices Exists Index Template API

IndicesExistsTemplate

Builder for the Indices Exists Template API

IndicesExistsType

Builder for the Indices Exists Type API

IndicesFlush

Builder for the Indices Flush API

IndicesFlushSynced

Builder for the Indices Flush Synced API

IndicesForcemerge

Builder for the Indices Forcemerge API

IndicesFreeze

Builder for the Indices Freeze API

IndicesGet

Builder for the Indices Get API

IndicesGetAlias

Builder for the Indices Get Alias API

IndicesGetDataStream

Builder for the Indices Get Data Stream API

IndicesGetFieldMapping

Builder for the Indices Get Field Mapping API

IndicesGetIndexTemplate

Builder for the Indices Get Index Template API

IndicesGetMapping

Builder for the Indices Get Mapping API

IndicesGetSettings

Builder for the Indices Get Settings API

IndicesGetTemplate

Builder for the Indices Get Template API

IndicesGetUpgrade

Builder for the Indices Get Upgrade API

IndicesMigrateToDataStream

Builder for the Indices Migrate To Data Stream API

IndicesOpen

Builder for the Indices Open API

IndicesPromoteDataStream

Builder for the Indices Promote Data Stream API

IndicesPutAlias

Builder for the Indices Put Alias API

IndicesPutIndexTemplate

Builder for the Indices Put Index Template API

IndicesPutMapping

Builder for the Indices Put Mapping API

IndicesPutSettings

Builder for the Indices Put Settings API

IndicesPutTemplate

Builder for the Indices Put Template API

IndicesRecovery

Builder for the Indices Recovery API

IndicesRefresh

Builder for the Indices Refresh API

IndicesReloadSearchAnalyzers

Builder for the Indices Reload Search Analyzers API

IndicesRollover

Builder for the Indices Rollover API

IndicesSegments

Builder for the Indices Segments API

IndicesShardStores

Builder for the Indices Shard Stores API

IndicesShrink

Builder for the Indices Shrink API

IndicesSimulateIndexTemplate

Builder for the Indices Simulate Index Template API

IndicesSimulateTemplate

Builder for the Indices Simulate Template API

IndicesSplit

Builder for the Indices Split API

IndicesStats

Builder for the Indices Stats API

IndicesUnfreeze

Builder for the Indices Unfreeze API

IndicesUpdateAliases

Builder for the Indices Update Aliases API

IndicesUpgrade

Builder for the Indices Upgrade API

IndicesValidateQuery

Builder for the Indices Validate Query API

Enums

IndicesAddBlockParts

API parts for the Indices Add Block API

IndicesAnalyzeParts

API parts for the Indices Analyze API

IndicesClearCacheParts

API parts for the Indices Clear Cache API

IndicesCloneParts

API parts for the Indices Clone API

IndicesCloseParts

API parts for the Indices Close API

IndicesCreateDataStreamParts

API parts for the Indices Create Data Stream API

IndicesCreateParts

API parts for the Indices Create API

IndicesDataStreamsStatsParts

API parts for the Indices Data Streams Stats API

IndicesDeleteAliasParts

API parts for the Indices Delete Alias API

IndicesDeleteDataStreamParts

API parts for the Indices Delete Data Stream API

IndicesDeleteIndexTemplateParts

API parts for the Indices Delete Index Template API

IndicesDeleteParts

API parts for the Indices Delete API

IndicesDeleteTemplateParts

API parts for the Indices Delete Template API

IndicesExistsAliasParts

API parts for the Indices Exists Alias API

IndicesExistsIndexTemplateParts

API parts for the Indices Exists Index Template API

IndicesExistsParts

API parts for the Indices Exists API

IndicesExistsTemplateParts

API parts for the Indices Exists Template API

IndicesExistsTypeParts

API parts for the Indices Exists Type API

IndicesFlushParts

API parts for the Indices Flush API

IndicesFlushSyncedParts

API parts for the Indices Flush Synced API

IndicesForcemergeParts

API parts for the Indices Forcemerge API

IndicesFreezeParts

API parts for the Indices Freeze API

IndicesGetAliasParts

API parts for the Indices Get Alias API

IndicesGetDataStreamParts

API parts for the Indices Get Data Stream API

IndicesGetFieldMappingParts

API parts for the Indices Get Field Mapping API

IndicesGetIndexTemplateParts

API parts for the Indices Get Index Template API

IndicesGetMappingParts

API parts for the Indices Get Mapping API

IndicesGetParts

API parts for the Indices Get API

IndicesGetSettingsParts

API parts for the Indices Get Settings API

IndicesGetTemplateParts

API parts for the Indices Get Template API

IndicesGetUpgradeParts

API parts for the Indices Get Upgrade API

IndicesMigrateToDataStreamParts

API parts for the Indices Migrate To Data Stream API

IndicesOpenParts

API parts for the Indices Open API

IndicesPromoteDataStreamParts

API parts for the Indices Promote Data Stream API

IndicesPutAliasParts

API parts for the Indices Put Alias API

IndicesPutIndexTemplateParts

API parts for the Indices Put Index Template API

IndicesPutMappingParts

API parts for the Indices Put Mapping API

IndicesPutSettingsParts

API parts for the Indices Put Settings API

IndicesPutTemplateParts

API parts for the Indices Put Template API

IndicesRecoveryParts

API parts for the Indices Recovery API

IndicesRefreshParts

API parts for the Indices Refresh API

IndicesReloadSearchAnalyzersParts

API parts for the Indices Reload Search Analyzers API

IndicesRolloverParts

API parts for the Indices Rollover API

IndicesSegmentsParts

API parts for the Indices Segments API

IndicesShardStoresParts

API parts for the Indices Shard Stores API

IndicesShrinkParts

API parts for the Indices Shrink API

IndicesSimulateIndexTemplateParts

API parts for the Indices Simulate Index Template API

IndicesSimulateTemplateParts

API parts for the Indices Simulate Template API

IndicesSplitParts

API parts for the Indices Split API

IndicesStatsParts

API parts for the Indices Stats API

IndicesUnfreezeParts

API parts for the Indices Unfreeze API

IndicesUpdateAliasesParts

API parts for the Indices Update Aliases API

IndicesUpgradeParts

API parts for the Indices Upgrade API

IndicesValidateQueryParts

API parts for the Indices Validate Query API