Struct elasticsearch::indices::Indices

pub struct Indices<'a> { /* private fields */ }

Namespace client for Indices APIs

Implementations

impl<'a> Indices<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Indices

pub fn transport(&self) -> &Transport

pub fn add_block<'b>(
&'a self,
parts: IndicesAddBlockParts<'b>
) -> IndicesAddBlock<'a, 'b, ()>

Indices Add Block API

Adds a block to an index.

pub fn analyze<'b>(
&'a self,
parts: IndicesAnalyzeParts<'b>
) -> IndicesAnalyze<'a, 'b, ()>

Indices Analyze API

Performs the analysis process on a text and return the tokens breakdown of the text.

pub fn clear_cache<'b>(
&'a self,
parts: IndicesClearCacheParts<'b>
) -> IndicesClearCache<'a, 'b, ()>

Indices Clear Cache API

Clears all or specific caches for one or more indices.

pub fn clone<'b>(
&'a self,
parts: IndicesCloneParts<'b>
) -> IndicesClone<'a, 'b, ()>

Indices Clone API

Clones an index

pub fn close<'b>(
&'a self,
parts: IndicesCloseParts<'b>
) -> IndicesClose<'a, 'b, ()>

Indices Close API

Closes an index.

pub fn create<'b>(
&'a self,
parts: IndicesCreateParts<'b>
) -> IndicesCreate<'a, 'b, ()>

Indices Create API

Creates an index with optional settings and mappings.

Examples

Create an index with a mapping

let client = Elasticsearch::default();
let response = client
    .indices()
    .create(IndicesCreateParts::Index("test_index"))
    .body(json!({
        "mappings" : {
            "properties" : {
                "field1" : { "type" : "text" }
            }
        }
    }))
    .send()
    .await?;
    

pub fn create_data_stream<'b>(
&'a self,
parts: IndicesCreateDataStreamParts<'b>
) -> IndicesCreateDataStream<'a, 'b, ()>

Indices Create Data Stream API

Creates a data stream

pub fn data_streams_stats<'b>(
&'a self,
parts: IndicesDataStreamsStatsParts<'b>
) -> IndicesDataStreamsStats<'a, 'b>

Indices Data Streams Stats API

Provides statistics on operations happening in a data stream.

pub fn delete<'b>(
&'a self,
parts: IndicesDeleteParts<'b>
) -> IndicesDelete<'a, 'b>

Indices Delete API

Deletes an index.

pub fn delete_alias<'b>(
&'a self,
parts: IndicesDeleteAliasParts<'b>
) -> IndicesDeleteAlias<'a, 'b>

Indices Delete Alias API

Deletes an alias.

pub fn delete_data_stream<'b>(
&'a self,
parts: IndicesDeleteDataStreamParts<'b>
) -> IndicesDeleteDataStream<'a, 'b>

Indices Delete Data Stream API

Deletes a data stream.

pub fn delete_index_template<'b>(
&'a self,
parts: IndicesDeleteIndexTemplateParts<'b>
) -> IndicesDeleteIndexTemplate<'a, 'b>

Indices Delete Index Template API

Deletes an index template.

pub fn delete_template<'b>(
&'a self,
parts: IndicesDeleteTemplateParts<'b>
) -> IndicesDeleteTemplate<'a, 'b>

Indices Delete Template API

Deletes an index template.

pub fn exists<'b>(
&'a self,
parts: IndicesExistsParts<'b>
) -> IndicesExists<'a, 'b>

Indices Exists API

Returns information about whether a particular index exists.

pub fn exists_alias<'b>(
&'a self,
parts: IndicesExistsAliasParts<'b>
) -> IndicesExistsAlias<'a, 'b>

Indices Exists Alias API

Returns information about whether a particular alias exists.

pub fn exists_index_template<'b>(
&'a self,
parts: IndicesExistsIndexTemplateParts<'b>
) -> IndicesExistsIndexTemplate<'a, 'b>

Indices Exists Index Template API

Returns information about whether a particular index template exists.

pub fn exists_template<'b>(
&'a self,
parts: IndicesExistsTemplateParts<'b>
) -> IndicesExistsTemplate<'a, 'b>

Indices Exists Template API

Returns information about whether a particular index template exists.

pub fn exists_type<'b>(
&'a self,
parts: IndicesExistsTypeParts<'b>
) -> IndicesExistsType<'a, 'b>

Indices Exists Type API

Returns information about whether a particular document type exists. (DEPRECATED)

pub fn flush<'b>(
&'a self,
parts: IndicesFlushParts<'b>
) -> IndicesFlush<'a, 'b, ()>

Indices Flush API

Performs the flush operation on one or more indices.

pub fn flush_synced<'b>(
&'a self,
parts: IndicesFlushSyncedParts<'b>
) -> IndicesFlushSynced<'a, 'b, ()>

Indices Flush Synced API

Performs a synced flush operation on one or more indices. Synced flush is deprecated and will be removed in 8.0. Use flush instead

pub fn forcemerge<'b>(
&'a self,
parts: IndicesForcemergeParts<'b>
) -> IndicesForcemerge<'a, 'b, ()>

Indices Forcemerge API

Performs the force merge operation on one or more indices.

pub fn freeze<'b>(
&'a self,
parts: IndicesFreezeParts<'b>
) -> IndicesFreeze<'a, 'b, ()>

Indices Freeze API

Freezes an index. A frozen index has almost no overhead on the cluster (except for maintaining its metadata in memory) and is read-only.

pub fn get<'b>(&'a self, parts: IndicesGetParts<'b>) -> IndicesGet<'a, 'b>

Indices Get API

Returns information about one or more indices.

pub fn get_alias<'b>(
&'a self,
parts: IndicesGetAliasParts<'b>
) -> IndicesGetAlias<'a, 'b>

Indices Get Alias API

Returns an alias.

pub fn get_data_stream<'b>(
&'a self,
parts: IndicesGetDataStreamParts<'b>
) -> IndicesGetDataStream<'a, 'b>

Indices Get Data Stream API

Returns data streams.

pub fn get_field_mapping<'b>(
&'a self,
parts: IndicesGetFieldMappingParts<'b>
) -> IndicesGetFieldMapping<'a, 'b>

Indices Get Field Mapping API

Returns mapping for one or more fields.

pub fn get_index_template<'b>(
&'a self,
parts: IndicesGetIndexTemplateParts<'b>
) -> IndicesGetIndexTemplate<'a, 'b>

Indices Get Index Template API

Returns an index template.

pub fn get_mapping<'b>(
&'a self,
parts: IndicesGetMappingParts<'b>
) -> IndicesGetMapping<'a, 'b>

Indices Get Mapping API

Returns mappings for one or more indices.

pub fn get_settings<'b>(
&'a self,
parts: IndicesGetSettingsParts<'b>
) -> IndicesGetSettings<'a, 'b>

Indices Get Settings API

Returns settings for one or more indices.

pub fn get_template<'b>(
&'a self,
parts: IndicesGetTemplateParts<'b>
) -> IndicesGetTemplate<'a, 'b>

Indices Get Template API

Returns an index template.

pub fn get_upgrade<'b>(
&'a self,
parts: IndicesGetUpgradeParts<'b>
) -> IndicesGetUpgrade<'a, 'b>

Indices Get Upgrade API

DEPRECATED Returns a progress status of current upgrade.

pub fn migrate_to_data_stream<'b>(
&'a self,
parts: IndicesMigrateToDataStreamParts<'b>
) -> IndicesMigrateToDataStream<'a, 'b, ()>

Indices Migrate To Data Stream API

Migrates an alias to a data stream

pub fn open<'b>(
&'a self,
parts: IndicesOpenParts<'b>
) -> IndicesOpen<'a, 'b, ()>

Indices Open API

Opens an index.

pub fn promote_data_stream<'b>(
&'a self,
parts: IndicesPromoteDataStreamParts<'b>
) -> IndicesPromoteDataStream<'a, 'b, ()>

Indices Promote Data Stream API

Promotes a data stream from a replicated data stream managed by CCR to a regular data stream

pub fn put_alias<'b>(
&'a self,
parts: IndicesPutAliasParts<'b>
) -> IndicesPutAlias<'a, 'b, ()>

Indices Put Alias API

Creates or updates an alias.

pub fn put_index_template<'b>(
&'a self,
parts: IndicesPutIndexTemplateParts<'b>
) -> IndicesPutIndexTemplate<'a, 'b, ()>

Indices Put Index Template API

Creates or updates an index template.

pub fn put_mapping<'b>(
&'a self,
parts: IndicesPutMappingParts<'b>
) -> IndicesPutMapping<'a, 'b, ()>

Indices Put Mapping API

Updates the index mappings.

Examples

Put a mapping into an existing index, assuming the index does not have a mapping, or that any properties specified do not conflict with existing properties

let client = Elasticsearch::default();
let response = client
    .indices()
    .put_mapping(IndicesPutMappingParts::Index(&["test_index"]))
    .body(json!({
        "properties" : {
            "field1" : { "type" : "text" }
        }
    }))
    .send()
    .await?;
    

pub fn put_settings<'b>(
&'a self,
parts: IndicesPutSettingsParts<'b>
) -> IndicesPutSettings<'a, 'b, ()>

Indices Put Settings API

Updates the index settings.

pub fn put_template<'b>(
&'a self,
parts: IndicesPutTemplateParts<'b>
) -> IndicesPutTemplate<'a, 'b, ()>

Indices Put Template API

Creates or updates an index template.

pub fn recovery<'b>(
&'a self,
parts: IndicesRecoveryParts<'b>
) -> IndicesRecovery<'a, 'b>

Indices Recovery API

Returns information about ongoing index shard recoveries.

pub fn refresh<'b>(
&'a self,
parts: IndicesRefreshParts<'b>
) -> IndicesRefresh<'a, 'b, ()>

Indices Refresh API

Performs the refresh operation in one or more indices.

pub fn reload_search_analyzers<'b>(
&'a self,
parts: IndicesReloadSearchAnalyzersParts<'b>
) -> IndicesReloadSearchAnalyzers<'a, 'b, ()>

Indices Reload Search Analyzers API

Reloads an index’s search analyzers and their resources.

pub fn rollover<'b>(
&'a self,
parts: IndicesRolloverParts<'b>
) -> IndicesRollover<'a, 'b, ()>

Indices Rollover API

Updates an alias to point to a new index when the existing index is considered to be too large or too old.

pub fn segments<'b>(
&'a self,
parts: IndicesSegmentsParts<'b>
) -> IndicesSegments<'a, 'b>

Indices Segments API

Provides low-level information about segments in a Lucene index.

pub fn shard_stores<'b>(
&'a self,
parts: IndicesShardStoresParts<'b>
) -> IndicesShardStores<'a, 'b>

Indices Shard Stores API

Provides store information for shard copies of indices.

pub fn shrink<'b>(
&'a self,
parts: IndicesShrinkParts<'b>
) -> IndicesShrink<'a, 'b, ()>

Indices Shrink API

Allow to shrink an existing index into a new index with fewer primary shards.

pub fn simulate_index_template<'b>(
&'a self,
parts: IndicesSimulateIndexTemplateParts<'b>
) -> IndicesSimulateIndexTemplate<'a, 'b, ()>

Indices Simulate Index Template API

Simulate matching the given index name against the index templates in the system

pub fn simulate_template<'b>(
&'a self,
parts: IndicesSimulateTemplateParts<'b>
) -> IndicesSimulateTemplate<'a, 'b, ()>

Indices Simulate Template API

Simulate resolving the given template name or body

pub fn split<'b>(
&'a self,
parts: IndicesSplitParts<'b>
) -> IndicesSplit<'a, 'b, ()>

Indices Split API

Allows you to split an existing index into a new index with more primary shards.

pub fn stats<'b>(&'a self, parts: IndicesStatsParts<'b>) -> IndicesStats<'a, 'b>

Indices Stats API

Provides statistics on operations happening in an index.

pub fn unfreeze<'b>(
&'a self,
parts: IndicesUnfreezeParts<'b>
) -> IndicesUnfreeze<'a, 'b, ()>

Indices Unfreeze API

Unfreezes an index. When a frozen index is unfrozen, the index goes through the normal recovery process and becomes writeable again.

pub fn update_aliases<'b>(&'a self) -> IndicesUpdateAliases<'a, 'b, ()>

Indices Update Aliases API

Updates index aliases.

pub fn upgrade<'b>(
&'a self,
parts: IndicesUpgradeParts<'b>
) -> IndicesUpgrade<'a, 'b, ()>

Indices Upgrade API

DEPRECATED Upgrades to the current version of Lucene.

pub fn validate_query<'b>(
&'a self,
parts: IndicesValidateQueryParts<'b>
) -> IndicesValidateQuery<'a, 'b, ()>

Indices Validate Query API

Allows a user to validate a potentially expensive query without executing it.