Module elasticsearch::ingest

Ingest APIs

Manage ingest pipelines. Ingest pipelines can be used on a node with the ingest role to pre-process documents before indexing, to apply transformations and enrich data. Transformations are performed by processors in the pipeline, and can include such operations as

add, remove and append fields within the document

point documents to the right time-based index based on a timestamp within the document

extract details from fields with known formats and add new fields with extracted data

and many more.

All nodes enable ingest by default, so any node can handle ingest tasks. Ingest pipelines can be conditionally executed, and failures within pipelines can be explicitly handled by defining processors to execute in the event of failure.

Structs

Ingest

Namespace client for Ingest APIs

IngestDeletePipeline

Builder for the Ingest Delete Pipeline API

IngestGeoIpStats

Builder for the Ingest Geo Ip Stats API

IngestGetPipeline

Builder for the Ingest Get Pipeline API

IngestProcessorGrok

Builder for the Ingest Processor Grok API

IngestPutPipeline

Builder for the Ingest Put Pipeline API

IngestSimulate

Builder for the Ingest Simulate API

Enums

IngestDeletePipelineParts

API parts for the Ingest Delete Pipeline API

IngestGeoIpStatsParts

API parts for the Ingest Geo Ip Stats API

IngestGetPipelineParts

API parts for the Ingest Get Pipeline API

IngestProcessorGrokParts

API parts for the Ingest Processor Grok API

IngestPutPipelineParts

API parts for the Ingest Put Pipeline API

IngestSimulateParts

API parts for the Ingest Simulate API