Module elasticsearch::ml

Machine Learning APIs

Perform machine learning anomaly detection activities.

Structs

Ml

Namespace client for Machine Learning APIs

MlCloseJob

Builder for the Ml Close Job API

MlDeleteCalendar

Builder for the Ml Delete Calendar API

MlDeleteCalendarEvent

Builder for the Ml Delete Calendar Event API

MlDeleteCalendarJob

Builder for the Ml Delete Calendar Job API

MlDeleteDataFrameAnalytics

Builder for the Ml Delete Data Frame Analytics API

MlDeleteDatafeed

Builder for the Ml Delete Datafeed API

MlDeleteExpiredData

Builder for the Ml Delete Expired Data API

MlDeleteFilter

Builder for the Ml Delete Filter API

MlDeleteForecast

Builder for the Ml Delete Forecast API

MlDeleteJob

Builder for the Ml Delete Job API

MlDeleteModelSnapshot

Builder for the Ml Delete Model Snapshot API

MlDeleteTrainedModel

Builder for the Ml Delete Trained Model API

MlDeleteTrainedModelAlias

Builder for the Ml Delete Trained Model Alias API

MlEstimateModelMemory

Builder for the Ml Estimate Model Memory API

MlEvaluateDataFrame

Builder for the Ml Evaluate Data Frame API

MlExplainDataFrameAnalytics

Builder for the Ml Explain Data Frame Analytics API

MlFlushJob

Builder for the Ml Flush Job API

MlForecast

Builder for the Ml Forecast API

MlGetBuckets

Builder for the Ml Get Buckets API

MlGetCalendarEvents

Builder for the Ml Get Calendar Events API

MlGetCalendars

Builder for the Ml Get Calendars API

MlGetCategories

Builder for the Ml Get Categories API

MlGetDataFrameAnalytics

Builder for the Ml Get Data Frame Analytics API

MlGetDataFrameAnalyticsStats

Builder for the Ml Get Data Frame Analytics Stats API

MlGetDatafeedStats

Builder for the Ml Get Datafeed Stats API

MlGetDatafeeds

Builder for the Ml Get Datafeeds API

MlGetFilters

Builder for the Ml Get Filters API

MlGetInfluencers

Builder for the Ml Get Influencers API

MlGetJobStats

Builder for the Ml Get Job Stats API

MlGetJobs

Builder for the Ml Get Jobs API

MlGetModelSnapshots

Builder for the Ml Get Model Snapshots API

MlGetOverallBuckets

Builder for the Ml Get Overall Buckets API

MlGetRecords

Builder for the Ml Get Records API

MlGetTrainedModels

Builder for the Ml Get Trained Models API

MlGetTrainedModelsStats

Builder for the Ml Get Trained Models Stats API

MlInfo

Builder for the Ml Info API

MlOpenJob

Builder for the Ml Open Job API

MlPostCalendarEvents

Builder for the Ml Post Calendar Events API

MlPostData

Builder for the Ml Post Data API

MlPreviewDataFrameAnalytics

Builder for the Ml Preview Data Frame Analytics API

MlPreviewDatafeed

Builder for the Ml Preview Datafeed API

MlPutCalendar

Builder for the Ml Put Calendar API

MlPutCalendarJob

Builder for the Ml Put Calendar Job API

MlPutDataFrameAnalytics

Builder for the Ml Put Data Frame Analytics API

MlPutDatafeed

Builder for the Ml Put Datafeed API

MlPutFilter

Builder for the Ml Put Filter API

MlPutJob

Builder for the Ml Put Job API

MlPutTrainedModel

Builder for the Ml Put Trained Model API

MlPutTrainedModelAlias

Builder for the Ml Put Trained Model Alias API

MlResetJob

Builder for the Ml Reset Job API

MlRevertModelSnapshot

Builder for the Ml Revert Model Snapshot API

MlSetUpgradeMode

Builder for the Ml Set Upgrade Mode API

MlStartDataFrameAnalytics

Builder for the Ml Start Data Frame Analytics API

MlStartDatafeed

Builder for the Ml Start Datafeed API

MlStopDataFrameAnalytics

Builder for the Ml Stop Data Frame Analytics API

MlStopDatafeed

Builder for the Ml Stop Datafeed API

MlUpdateDataFrameAnalytics

Builder for the Ml Update Data Frame Analytics API

MlUpdateDatafeed

Builder for the Ml Update Datafeed API

MlUpdateFilter

Builder for the Ml Update Filter API

MlUpdateJob

Builder for the Ml Update Job API

MlUpdateModelSnapshot

Builder for the Ml Update Model Snapshot API

MlUpgradeJobSnapshot

Builder for the Ml Upgrade Job Snapshot API

MlValidate

Builder for the Ml Validate API

MlValidateDetector

Builder for the Ml Validate Detector API

Enums

MlCloseJobParts

API parts for the Ml Close Job API

MlDeleteCalendarEventParts

API parts for the Ml Delete Calendar Event API

MlDeleteCalendarJobParts

API parts for the Ml Delete Calendar Job API

MlDeleteCalendarParts

API parts for the Ml Delete Calendar API

MlDeleteDataFrameAnalyticsParts

API parts for the Ml Delete Data Frame Analytics API

MlDeleteDatafeedParts

API parts for the Ml Delete Datafeed API

MlDeleteExpiredDataParts

API parts for the Ml Delete Expired Data API

MlDeleteFilterParts

API parts for the Ml Delete Filter API

MlDeleteForecastParts

API parts for the Ml Delete Forecast API

MlDeleteJobParts

API parts for the Ml Delete Job API

MlDeleteModelSnapshotParts

API parts for the Ml Delete Model Snapshot API

MlDeleteTrainedModelAliasParts

API parts for the Ml Delete Trained Model Alias API

MlDeleteTrainedModelParts

API parts for the Ml Delete Trained Model API

MlEstimateModelMemoryParts

API parts for the Ml Estimate Model Memory API

MlEvaluateDataFrameParts

API parts for the Ml Evaluate Data Frame API

MlExplainDataFrameAnalyticsParts

API parts for the Ml Explain Data Frame Analytics API

MlFlushJobParts

API parts for the Ml Flush Job API

MlForecastParts

API parts for the Ml Forecast API

MlGetBucketsParts

API parts for the Ml Get Buckets API

MlGetCalendarEventsParts

API parts for the Ml Get Calendar Events API

MlGetCalendarsParts

API parts for the Ml Get Calendars API

MlGetCategoriesParts

API parts for the Ml Get Categories API

MlGetDataFrameAnalyticsParts

API parts for the Ml Get Data Frame Analytics API

MlGetDataFrameAnalyticsStatsParts

API parts for the Ml Get Data Frame Analytics Stats API

MlGetDatafeedStatsParts

API parts for the Ml Get Datafeed Stats API

MlGetDatafeedsParts

API parts for the Ml Get Datafeeds API

MlGetFiltersParts

API parts for the Ml Get Filters API

MlGetInfluencersParts

API parts for the Ml Get Influencers API

MlGetJobStatsParts

API parts for the Ml Get Job Stats API

MlGetJobsParts

API parts for the Ml Get Jobs API

MlGetModelSnapshotsParts

API parts for the Ml Get Model Snapshots API

MlGetOverallBucketsParts

API parts for the Ml Get Overall Buckets API

MlGetRecordsParts

API parts for the Ml Get Records API

MlGetTrainedModelsParts

API parts for the Ml Get Trained Models API

MlGetTrainedModelsStatsParts

API parts for the Ml Get Trained Models Stats API

MlInfoParts

API parts for the Ml Info API

MlOpenJobParts

API parts for the Ml Open Job API

MlPostCalendarEventsParts

API parts for the Ml Post Calendar Events API

MlPostDataParts

API parts for the Ml Post Data API

MlPreviewDataFrameAnalyticsParts

API parts for the Ml Preview Data Frame Analytics API

MlPreviewDatafeedParts

API parts for the Ml Preview Datafeed API

MlPutCalendarJobParts

API parts for the Ml Put Calendar Job API

MlPutCalendarParts

API parts for the Ml Put Calendar API

MlPutDataFrameAnalyticsParts

API parts for the Ml Put Data Frame Analytics API

MlPutDatafeedParts

API parts for the Ml Put Datafeed API

MlPutFilterParts

API parts for the Ml Put Filter API

MlPutJobParts

API parts for the Ml Put Job API

MlPutTrainedModelAliasParts

API parts for the Ml Put Trained Model Alias API

MlPutTrainedModelParts

API parts for the Ml Put Trained Model API

MlResetJobParts

API parts for the Ml Reset Job API

MlRevertModelSnapshotParts

API parts for the Ml Revert Model Snapshot API

MlSetUpgradeModeParts

API parts for the Ml Set Upgrade Mode API

MlStartDataFrameAnalyticsParts

API parts for the Ml Start Data Frame Analytics API

MlStartDatafeedParts

API parts for the Ml Start Datafeed API

MlStopDataFrameAnalyticsParts

API parts for the Ml Stop Data Frame Analytics API

MlStopDatafeedParts

API parts for the Ml Stop Datafeed API

MlUpdateDataFrameAnalyticsParts

API parts for the Ml Update Data Frame Analytics API

MlUpdateDatafeedParts

API parts for the Ml Update Datafeed API

MlUpdateFilterParts

API parts for the Ml Update Filter API

MlUpdateJobParts

API parts for the Ml Update Job API

MlUpdateModelSnapshotParts

API parts for the Ml Update Model Snapshot API

MlUpgradeJobSnapshotParts

API parts for the Ml Upgrade Job Snapshot API

MlValidateDetectorParts

API parts for the Ml Validate Detector API

MlValidateParts

API parts for the Ml Validate API