Struct elasticsearch::ml::Ml

pub struct Ml<'a> { /* private fields */ }

Namespace client for Machine Learning APIs

Implementations

impl<'a> Ml<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Ml

pub fn transport(&self) -> &Transport

pub fn close_job<'b>(
&'a self,
parts: MlCloseJobParts<'b>
) -> MlCloseJob<'a, 'b, ()>

Ml Close Job API

Closes one or more anomaly detection jobs. A job can be opened and closed multiple times throughout its lifecycle.

pub fn delete_calendar<'b>(
&'a self,
parts: MlDeleteCalendarParts<'b>
) -> MlDeleteCalendar<'a, 'b>

Ml Delete Calendar API

Deletes a calendar.

pub fn delete_calendar_event<'b>(
&'a self,
parts: MlDeleteCalendarEventParts<'b>
) -> MlDeleteCalendarEvent<'a, 'b>

Ml Delete Calendar Event API

Deletes scheduled events from a calendar.

pub fn delete_calendar_job<'b>(
&'a self,
parts: MlDeleteCalendarJobParts<'b>
) -> MlDeleteCalendarJob<'a, 'b>

Ml Delete Calendar Job API

Deletes anomaly detection jobs from a calendar.

pub fn delete_data_frame_analytics<'b>(
&'a self,
parts: MlDeleteDataFrameAnalyticsParts<'b>
) -> MlDeleteDataFrameAnalytics<'a, 'b>

Ml Delete Data Frame Analytics API

Deletes an existing data frame analytics job.

pub fn delete_datafeed<'b>(
&'a self,
parts: MlDeleteDatafeedParts<'b>
) -> MlDeleteDatafeed<'a, 'b>

Ml Delete Datafeed API

Deletes an existing datafeed.

pub fn delete_expired_data<'b>(
&'a self,
parts: MlDeleteExpiredDataParts<'b>
) -> MlDeleteExpiredData<'a, 'b, ()>

Ml Delete Expired Data API

Deletes expired and unused machine learning data.

pub fn delete_filter<'b>(
&'a self,
parts: MlDeleteFilterParts<'b>
) -> MlDeleteFilter<'a, 'b>

Ml Delete Filter API

Deletes a filter.

pub fn delete_forecast<'b>(
&'a self,
parts: MlDeleteForecastParts<'b>
) -> MlDeleteForecast<'a, 'b>

Ml Delete Forecast API

Deletes forecasts from a machine learning job.

pub fn delete_job<'b>(
&'a self,
parts: MlDeleteJobParts<'b>
) -> MlDeleteJob<'a, 'b>

Ml Delete Job API

Deletes an existing anomaly detection job.

pub fn delete_model_snapshot<'b>(
&'a self,
parts: MlDeleteModelSnapshotParts<'b>
) -> MlDeleteModelSnapshot<'a, 'b>

Ml Delete Model Snapshot API

Deletes an existing model snapshot.

pub fn delete_trained_model<'b>(
&'a self,
parts: MlDeleteTrainedModelParts<'b>
) -> MlDeleteTrainedModel<'a, 'b>

Ml Delete Trained Model API

Deletes an existing trained inference model that is currently not referenced by an ingest pipeline.

pub fn delete_trained_model_alias<'b>(
&'a self,
parts: MlDeleteTrainedModelAliasParts<'b>
) -> MlDeleteTrainedModelAlias<'a, 'b>

Ml Delete Trained Model Alias API

Deletes a model alias that refers to the trained model

pub fn estimate_model_memory<'b>(&'a self) -> MlEstimateModelMemory<'a, 'b, ()>

Ml Estimate Model Memory API

Estimates the model memory

pub fn evaluate_data_frame<'b>(&'a self) -> MlEvaluateDataFrame<'a, 'b, ()>

Ml Evaluate Data Frame API

Evaluates the data frame analytics for an annotated index.

pub fn explain_data_frame_analytics<'b>(
&'a self,
parts: MlExplainDataFrameAnalyticsParts<'b>
) -> MlExplainDataFrameAnalytics<'a, 'b, ()>

Ml Explain Data Frame Analytics API

Explains a data frame analytics config.

pub fn flush_job<'b>(
&'a self,
parts: MlFlushJobParts<'b>
) -> MlFlushJob<'a, 'b, ()>

Ml Flush Job API

Forces any buffered data to be processed by the job.

pub fn forecast<'b>(
&'a self,
parts: MlForecastParts<'b>
) -> MlForecast<'a, 'b, ()>

Ml Forecast API

Predicts the future behavior of a time series by using its historical behavior.

pub fn get_buckets<'b>(
&'a self,
parts: MlGetBucketsParts<'b>
) -> MlGetBuckets<'a, 'b, ()>

Ml Get Buckets API

Retrieves anomaly detection job results for one or more buckets.

pub fn get_calendar_events<'b>(
&'a self,
parts: MlGetCalendarEventsParts<'b>
) -> MlGetCalendarEvents<'a, 'b>

Ml Get Calendar Events API

Retrieves information about the scheduled events in calendars.

pub fn get_calendars<'b>(
&'a self,
parts: MlGetCalendarsParts<'b>
) -> MlGetCalendars<'a, 'b, ()>

Ml Get Calendars API

Retrieves configuration information for calendars.

pub fn get_categories<'b>(
&'a self,
parts: MlGetCategoriesParts<'b>
) -> MlGetCategories<'a, 'b, ()>

Ml Get Categories API

Retrieves anomaly detection job results for one or more categories.

pub fn get_data_frame_analytics<'b>(
&'a self,
parts: MlGetDataFrameAnalyticsParts<'b>
) -> MlGetDataFrameAnalytics<'a, 'b>

Ml Get Data Frame Analytics API

Retrieves configuration information for data frame analytics jobs.

pub fn get_data_frame_analytics_stats<'b>(
&'a self,
parts: MlGetDataFrameAnalyticsStatsParts<'b>
) -> MlGetDataFrameAnalyticsStats<'a, 'b>

Ml Get Data Frame Analytics Stats API

Retrieves usage information for data frame analytics jobs.

pub fn get_datafeed_stats<'b>(
&'a self,
parts: MlGetDatafeedStatsParts<'b>
) -> MlGetDatafeedStats<'a, 'b>

Ml Get Datafeed Stats API

Retrieves usage information for datafeeds.

pub fn get_datafeeds<'b>(
&'a self,
parts: MlGetDatafeedsParts<'b>
) -> MlGetDatafeeds<'a, 'b>

Ml Get Datafeeds API

Retrieves configuration information for datafeeds.

pub fn get_filters<'b>(
&'a self,
parts: MlGetFiltersParts<'b>
) -> MlGetFilters<'a, 'b>

Ml Get Filters API

Retrieves filters.

pub fn get_influencers<'b>(
&'a self,
parts: MlGetInfluencersParts<'b>
) -> MlGetInfluencers<'a, 'b, ()>

Ml Get Influencers API

Retrieves anomaly detection job results for one or more influencers.

pub fn get_job_stats<'b>(
&'a self,
parts: MlGetJobStatsParts<'b>
) -> MlGetJobStats<'a, 'b>

Ml Get Job Stats API

Retrieves usage information for anomaly detection jobs.

pub fn get_jobs<'b>(&'a self, parts: MlGetJobsParts<'b>) -> MlGetJobs<'a, 'b>

Ml Get Jobs API

Retrieves configuration information for anomaly detection jobs.

pub fn get_model_snapshots<'b>(
&'a self,
parts: MlGetModelSnapshotsParts<'b>
) -> MlGetModelSnapshots<'a, 'b, ()>

Ml Get Model Snapshots API

Retrieves information about model snapshots.

pub fn get_overall_buckets<'b>(
&'a self,
parts: MlGetOverallBucketsParts<'b>
) -> MlGetOverallBuckets<'a, 'b, ()>

Ml Get Overall Buckets API

Retrieves overall bucket results that summarize the bucket results of multiple anomaly detection jobs.

pub fn get_records<'b>(
&'a self,
parts: MlGetRecordsParts<'b>
) -> MlGetRecords<'a, 'b, ()>

Ml Get Records API

Retrieves anomaly records for an anomaly detection job.

pub fn get_trained_models<'b>(
&'a self,
parts: MlGetTrainedModelsParts<'b>
) -> MlGetTrainedModels<'a, 'b>

Ml Get Trained Models API

Retrieves configuration information for a trained inference model.

pub fn get_trained_models_stats<'b>(
&'a self,
parts: MlGetTrainedModelsStatsParts<'b>
) -> MlGetTrainedModelsStats<'a, 'b>

Ml Get Trained Models Stats API

Retrieves usage information for trained inference models.

pub fn info<'b>(&'a self) -> MlInfo<'a, 'b>

Ml Info API

Returns defaults and limits used by machine learning.

pub fn open_job<'b>(
&'a self,
parts: MlOpenJobParts<'b>
) -> MlOpenJob<'a, 'b, ()>

Ml Open Job API

Opens one or more anomaly detection jobs.

pub fn post_calendar_events<'b>(
&'a self,
parts: MlPostCalendarEventsParts<'b>
) -> MlPostCalendarEvents<'a, 'b, ()>

Ml Post Calendar Events API

Posts scheduled events in a calendar.

pub fn post_data<'b>(
&'a self,
parts: MlPostDataParts<'b>
) -> MlPostData<'a, 'b, ()>

Ml Post Data API

Sends data to an anomaly detection job for analysis.

pub fn preview_data_frame_analytics<'b>(
&'a self,
parts: MlPreviewDataFrameAnalyticsParts<'b>
) -> MlPreviewDataFrameAnalytics<'a, 'b, ()>

Ml Preview Data Frame Analytics API

Previews that will be analyzed given a data frame analytics config.

pub fn preview_datafeed<'b>(
&'a self,
parts: MlPreviewDatafeedParts<'b>
) -> MlPreviewDatafeed<'a, 'b, ()>

Ml Preview Datafeed API

Previews a datafeed.

pub fn put_calendar<'b>(
&'a self,
parts: MlPutCalendarParts<'b>
) -> MlPutCalendar<'a, 'b, ()>

Ml Put Calendar API

Instantiates a calendar.

pub fn put_calendar_job<'b>(
&'a self,
parts: MlPutCalendarJobParts<'b>
) -> MlPutCalendarJob<'a, 'b, ()>

Ml Put Calendar Job API

Adds an anomaly detection job to a calendar.

pub fn put_data_frame_analytics<'b>(
&'a self,
parts: MlPutDataFrameAnalyticsParts<'b>
) -> MlPutDataFrameAnalytics<'a, 'b, ()>

Ml Put Data Frame Analytics API

Instantiates a data frame analytics job.

pub fn put_datafeed<'b>(
&'a self,
parts: MlPutDatafeedParts<'b>
) -> MlPutDatafeed<'a, 'b, ()>

Ml Put Datafeed API

Instantiates a datafeed.

pub fn put_filter<'b>(
&'a self,
parts: MlPutFilterParts<'b>
) -> MlPutFilter<'a, 'b, ()>

Ml Put Filter API

Instantiates a filter.

pub fn put_job<'b>(&'a self, parts: MlPutJobParts<'b>) -> MlPutJob<'a, 'b, ()>

Ml Put Job API

Instantiates an anomaly detection job.

pub fn put_trained_model<'b>(
&'a self,
parts: MlPutTrainedModelParts<'b>
) -> MlPutTrainedModel<'a, 'b, ()>

Ml Put Trained Model API

Creates an inference trained model.

pub fn put_trained_model_alias<'b>(
&'a self,
parts: MlPutTrainedModelAliasParts<'b>
) -> MlPutTrainedModelAlias<'a, 'b, ()>

Ml Put Trained Model Alias API

Creates a new model alias (or reassigns an existing one) to refer to the trained model

pub fn reset_job<'b>(
&'a self,
parts: MlResetJobParts<'b>
) -> MlResetJob<'a, 'b, ()>

Ml Reset Job API

Resets an existing anomaly detection job.

pub fn revert_model_snapshot<'b>(
&'a self,
parts: MlRevertModelSnapshotParts<'b>
) -> MlRevertModelSnapshot<'a, 'b, ()>

Ml Revert Model Snapshot API

Reverts to a specific snapshot.

pub fn set_upgrade_mode<'b>(&'a self) -> MlSetUpgradeMode<'a, 'b, ()>

Ml Set Upgrade Mode API

Sets a cluster wide upgrade_mode setting that prepares machine learning indices for an upgrade.

pub fn start_data_frame_analytics<'b>(
&'a self,
parts: MlStartDataFrameAnalyticsParts<'b>
) -> MlStartDataFrameAnalytics<'a, 'b, ()>

Ml Start Data Frame Analytics API

Starts a data frame analytics job.

pub fn start_datafeed<'b>(
&'a self,
parts: MlStartDatafeedParts<'b>
) -> MlStartDatafeed<'a, 'b, ()>

Ml Start Datafeed API

Starts one or more datafeeds.

pub fn stop_data_frame_analytics<'b>(
&'a self,
parts: MlStopDataFrameAnalyticsParts<'b>
) -> MlStopDataFrameAnalytics<'a, 'b, ()>

Ml Stop Data Frame Analytics API

Stops one or more data frame analytics jobs.

pub fn stop_datafeed<'b>(
&'a self,
parts: MlStopDatafeedParts<'b>
) -> MlStopDatafeed<'a, 'b, ()>

Ml Stop Datafeed API

Stops one or more datafeeds.

pub fn update_data_frame_analytics<'b>(
&'a self,
parts: MlUpdateDataFrameAnalyticsParts<'b>
) -> MlUpdateDataFrameAnalytics<'a, 'b, ()>

Ml Update Data Frame Analytics API

Updates certain properties of a data frame analytics job.

pub fn update_datafeed<'b>(
&'a self,
parts: MlUpdateDatafeedParts<'b>
) -> MlUpdateDatafeed<'a, 'b, ()>

Ml Update Datafeed API

Updates certain properties of a datafeed.

pub fn update_filter<'b>(
&'a self,
parts: MlUpdateFilterParts<'b>
) -> MlUpdateFilter<'a, 'b, ()>

Ml Update Filter API

Updates the description of a filter, adds items, or removes items.

pub fn update_job<'b>(
&'a self,
parts: MlUpdateJobParts<'b>
) -> MlUpdateJob<'a, 'b, ()>

Ml Update Job API

Updates certain properties of an anomaly detection job.

pub fn update_model_snapshot<'b>(
&'a self,
parts: MlUpdateModelSnapshotParts<'b>
) -> MlUpdateModelSnapshot<'a, 'b, ()>

Ml Update Model Snapshot API

Updates certain properties of a snapshot.

pub fn upgrade_job_snapshot<'b>(
&'a self,
parts: MlUpgradeJobSnapshotParts<'b>
) -> MlUpgradeJobSnapshot<'a, 'b, ()>

Ml Upgrade Job Snapshot API

Upgrades a given job snapshot to the current major version.

pub fn validate<'b>(&'a self) -> MlValidate<'a, 'b, ()>

Ml Validate API

Validates an anomaly detection job.

pub fn validate_detector<'b>(&'a self) -> MlValidateDetector<'a, 'b, ()>

Ml Validate Detector API

Validates an anomaly detection detector.