Enum elasticsearch::params::SourceFilter

pub enum SourceFilter {
    Enable(bool),
    Include(String),
    Includes(Vec<String>),
    IncludesExcludes {
        includes: Vec<String>,
        excludes: Vec<String>,
    },
}

Control how the _source field is returned with every hit.

By default operations return the contents of the _source field unless you have used the stored_fields parameter or if the _source field is disabled.

Variants

Enable(bool)

Tuple Fields

0: bool

Whether _source retrieval should be enabled (true) or disabled (false)

Include(String)

Tuple Fields

0: String

A wildcard pattern to control what parts of _source should be returned

Includes(Vec<String>)

Tuple Fields

0: Vec<String>

A collection of wildcard patterns to control what parts of _source should be returned

IncludesExcludes

Fields

includes: Vec<String>

excludes: Vec<String>

A collection of wildcard patterns to control what parts of _source should and should not be returned