Module elasticsearch::security

Security APIs

Perform security related activities, including

Manage users

Manage application privileges

Manage roles

Manage role mappings

Manage API keys

Manage Bearer tokens

Authenticate users against an OpenID Connect or SAML authentication realm when using a custom web application other than Kibana

Structs

Security

Namespace client for Security APIs

SecurityAuthenticate

Builder for the Security Authenticate API

SecurityChangePassword

Builder for the Security Change Password API

SecurityClearApiKeyCache

Builder for the Security Clear Api Key Cache API

SecurityClearCachedPrivileges

Builder for the Security Clear Cached Privileges API

SecurityClearCachedRealms

Builder for the Security Clear Cached Realms API

SecurityClearCachedRoles

Builder for the Security Clear Cached Roles API

SecurityCreateApiKey

Builder for the Security Create Api Key API

SecurityDeletePrivileges

Builder for the Security Delete Privileges API

SecurityDeleteRole

Builder for the Security Delete Role API

SecurityDeleteRoleMapping

Builder for the Security Delete Role Mapping API

SecurityDeleteUser

Builder for the Security Delete User API

SecurityDisableUser

Builder for the Security Disable User API

SecurityEnableUser

Builder for the Security Enable User API

SecurityGetApiKey

Builder for the Security Get Api Key API

SecurityGetBuiltinPrivileges

Builder for the Security Get Builtin Privileges API

SecurityGetPrivileges

Builder for the Security Get Privileges API

SecurityGetRole

Builder for the Security Get Role API

SecurityGetRoleMapping

Builder for the Security Get Role Mapping API

SecurityGetToken

Builder for the Security Get Token API

SecurityGetUser

Builder for the Security Get User API

SecurityGetUserPrivileges

Builder for the Security Get User Privileges API

SecurityGrantApiKey

Builder for the Security Grant Api Key API

SecurityHasPrivileges

Builder for the Security Has Privileges API

SecurityInvalidateApiKey

Builder for the Security Invalidate Api Key API

SecurityInvalidateToken

Builder for the Security Invalidate Token API

SecurityPutPrivileges

Builder for the Security Put Privileges API

SecurityPutRole

Builder for the Security Put Role API

SecurityPutRoleMapping

Builder for the Security Put Role Mapping API

SecurityPutUser

Builder for the Security Put User API

SecuritySamlAuthenticate

Builder for the Security Saml Authenticate API

SecuritySamlCompleteLogout

Builder for the Security Saml Complete Logout API

SecuritySamlInvalidate

Builder for the Security Saml Invalidate API

SecuritySamlLogout

Builder for the Security Saml Logout API

SecuritySamlPrepareAuthentication

Builder for the Security Saml Prepare Authentication API

SecuritySamlServiceProviderMetadata

Builder for the Security Saml Service Provider Metadata API

Enums

SecurityAuthenticateParts

API parts for the Security Authenticate API

SecurityChangePasswordParts

API parts for the Security Change Password API

SecurityClearApiKeyCacheParts

API parts for the Security Clear Api Key Cache API

SecurityClearCachedPrivilegesParts

API parts for the Security Clear Cached Privileges API

SecurityClearCachedRealmsParts

API parts for the Security Clear Cached Realms API

SecurityClearCachedRolesParts

API parts for the Security Clear Cached Roles API

SecurityCreateApiKeyParts

API parts for the Security Create Api Key API

SecurityDeletePrivilegesParts

API parts for the Security Delete Privileges API

SecurityDeleteRoleMappingParts

API parts for the Security Delete Role Mapping API

SecurityDeleteRoleParts

API parts for the Security Delete Role API

SecurityDeleteUserParts

API parts for the Security Delete User API

SecurityDisableUserParts

API parts for the Security Disable User API

SecurityEnableUserParts

API parts for the Security Enable User API

SecurityGetApiKeyParts

API parts for the Security Get Api Key API

SecurityGetBuiltinPrivilegesParts

API parts for the Security Get Builtin Privileges API

SecurityGetPrivilegesParts

API parts for the Security Get Privileges API

SecurityGetRoleMappingParts

API parts for the Security Get Role Mapping API

SecurityGetRoleParts

API parts for the Security Get Role API

SecurityGetTokenParts

API parts for the Security Get Token API

SecurityGetUserParts

API parts for the Security Get User API

SecurityGetUserPrivilegesParts

API parts for the Security Get User Privileges API

SecurityGrantApiKeyParts

API parts for the Security Grant Api Key API

SecurityHasPrivilegesParts

API parts for the Security Has Privileges API

SecurityInvalidateApiKeyParts

API parts for the Security Invalidate Api Key API

SecurityInvalidateTokenParts

API parts for the Security Invalidate Token API

SecurityPutPrivilegesParts

API parts for the Security Put Privileges API

SecurityPutRoleMappingParts

API parts for the Security Put Role Mapping API

SecurityPutRoleParts

API parts for the Security Put Role API

SecurityPutUserParts

API parts for the Security Put User API

SecuritySamlAuthenticateParts

API parts for the Security Saml Authenticate API

SecuritySamlCompleteLogoutParts

API parts for the Security Saml Complete Logout API

SecuritySamlInvalidateParts

API parts for the Security Saml Invalidate API

SecuritySamlLogoutParts

API parts for the Security Saml Logout API

SecuritySamlPrepareAuthenticationParts

API parts for the Security Saml Prepare Authentication API

SecuritySamlServiceProviderMetadataParts

API parts for the Security Saml Service Provider Metadata API