Struct elasticsearch::security::Security

pub struct Security<'a> { /* private fields */ }

Namespace client for Security APIs

Implementations

impl<'a> Security<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Security

pub fn transport(&self) -> &Transport

pub fn authenticate<'b>(&'a self) -> SecurityAuthenticate<'a, 'b>

Security Authenticate API

Enables authentication as a user and retrieve information about the authenticated user.

pub fn change_password<'b>(
&'a self,
parts: SecurityChangePasswordParts<'b>
) -> SecurityChangePassword<'a, 'b, ()>

Security Change Password API

Changes the passwords of users in the native realm and built-in users.

pub fn clear_api_key_cache<'b>(
&'a self,
parts: SecurityClearApiKeyCacheParts<'b>
) -> SecurityClearApiKeyCache<'a, 'b, ()>

Security Clear Api Key Cache API

Clear a subset or all entries from the API key cache.

pub fn clear_cached_privileges<'b>(
&'a self,
parts: SecurityClearCachedPrivilegesParts<'b>
) -> SecurityClearCachedPrivileges<'a, 'b, ()>

Security Clear Cached Privileges API

Evicts application privileges from the native application privileges cache.

pub fn clear_cached_realms<'b>(
&'a self,
parts: SecurityClearCachedRealmsParts<'b>
) -> SecurityClearCachedRealms<'a, 'b, ()>

Security Clear Cached Realms API

Evicts users from the user cache. Can completely clear the cache or evict specific users.

pub fn clear_cached_roles<'b>(
&'a self,
parts: SecurityClearCachedRolesParts<'b>
) -> SecurityClearCachedRoles<'a, 'b, ()>

Security Clear Cached Roles API

Evicts roles from the native role cache.

pub fn create_api_key<'b>(&'a self) -> SecurityCreateApiKey<'a, 'b, ()>

Security Create Api Key API

Creates an API key for access without requiring basic authentication.

pub fn delete_privileges<'b>(
&'a self,
parts: SecurityDeletePrivilegesParts<'b>
) -> SecurityDeletePrivileges<'a, 'b>

Security Delete Privileges API

Removes application privileges.

pub fn delete_role<'b>(
&'a self,
parts: SecurityDeleteRoleParts<'b>
) -> SecurityDeleteRole<'a, 'b>

Security Delete Role API

Removes roles in the native realm.

pub fn delete_role_mapping<'b>(
&'a self,
parts: SecurityDeleteRoleMappingParts<'b>
) -> SecurityDeleteRoleMapping<'a, 'b>

Security Delete Role Mapping API

Removes role mappings.

pub fn delete_user<'b>(
&'a self,
parts: SecurityDeleteUserParts<'b>
) -> SecurityDeleteUser<'a, 'b>

Security Delete User API

Deletes users from the native realm.

pub fn disable_user<'b>(
&'a self,
parts: SecurityDisableUserParts<'b>
) -> SecurityDisableUser<'a, 'b, ()>

Security Disable User API

Disables users in the native realm.

pub fn enable_user<'b>(
&'a self,
parts: SecurityEnableUserParts<'b>
) -> SecurityEnableUser<'a, 'b, ()>

Security Enable User API

Enables users in the native realm.

pub fn get_api_key<'b>(&'a self) -> SecurityGetApiKey<'a, 'b>

Security Get Api Key API

Retrieves information for one or more API keys.

pub fn get_builtin_privileges<'b>(
&'a self
) -> SecurityGetBuiltinPrivileges<'a, 'b>

Security Get Builtin Privileges API

Retrieves the list of cluster privileges and index privileges that are available in this version of Elasticsearch.

pub fn get_privileges<'b>(
&'a self,
parts: SecurityGetPrivilegesParts<'b>
) -> SecurityGetPrivileges<'a, 'b>

Security Get Privileges API

Retrieves application privileges.

pub fn get_role<'b>(
&'a self,
parts: SecurityGetRoleParts<'b>
) -> SecurityGetRole<'a, 'b>

Security Get Role API

Retrieves roles in the native realm.

pub fn get_role_mapping<'b>(
&'a self,
parts: SecurityGetRoleMappingParts<'b>
) -> SecurityGetRoleMapping<'a, 'b>

Security Get Role Mapping API

Retrieves role mappings.

pub fn get_token<'b>(&'a self) -> SecurityGetToken<'a, 'b, ()>

Security Get Token API

Creates a bearer token for access without requiring basic authentication.

pub fn get_user<'b>(
&'a self,
parts: SecurityGetUserParts<'b>
) -> SecurityGetUser<'a, 'b>

Security Get User API

Retrieves information about users in the native realm and built-in users.

pub fn get_user_privileges<'b>(&'a self) -> SecurityGetUserPrivileges<'a, 'b>

Security Get User Privileges API

Retrieves security privileges for the logged in user.

pub fn grant_api_key<'b>(&'a self) -> SecurityGrantApiKey<'a, 'b, ()>

Security Grant Api Key API

Creates an API key on behalf of another user.

pub fn has_privileges<'b>(
&'a self,
parts: SecurityHasPrivilegesParts<'b>
) -> SecurityHasPrivileges<'a, 'b, ()>

Security Has Privileges API

Determines whether the specified user has a specified list of privileges.

pub fn invalidate_api_key<'b>(&'a self) -> SecurityInvalidateApiKey<'a, 'b, ()>

Security Invalidate Api Key API

Invalidates one or more API keys.

pub fn invalidate_token<'b>(&'a self) -> SecurityInvalidateToken<'a, 'b, ()>

Security Invalidate Token API

Invalidates one or more access tokens or refresh tokens.

pub fn put_privileges<'b>(&'a self) -> SecurityPutPrivileges<'a, 'b, ()>

Security Put Privileges API

Adds or updates application privileges.

pub fn put_role<'b>(
&'a self,
parts: SecurityPutRoleParts<'b>
) -> SecurityPutRole<'a, 'b, ()>

Security Put Role API

Adds and updates roles in the native realm.

pub fn put_role_mapping<'b>(
&'a self,
parts: SecurityPutRoleMappingParts<'b>
) -> SecurityPutRoleMapping<'a, 'b, ()>

Security Put Role Mapping API

Creates and updates role mappings.

pub fn put_user<'b>(
&'a self,
parts: SecurityPutUserParts<'b>
) -> SecurityPutUser<'a, 'b, ()>

Security Put User API

Adds and updates users in the native realm. These users are commonly referred to as native users.

pub fn saml_authenticate<'b>(&'a self) -> SecuritySamlAuthenticate<'a, 'b, ()>

Security Saml Authenticate API

Exchanges a SAML Response message for an Elasticsearch access token and refresh token pair

pub fn saml_complete_logout<'b>(
&'a self
) -> SecuritySamlCompleteLogout<'a, 'b, ()>

Security Saml Complete Logout API

Verifies the logout response sent from the SAML IdP

pub fn saml_invalidate<'b>(&'a self) -> SecuritySamlInvalidate<'a, 'b, ()>

Security Saml Invalidate API

Consumes a SAML LogoutRequest

pub fn saml_logout<'b>(&'a self) -> SecuritySamlLogout<'a, 'b, ()>

Security Saml Logout API

Invalidates an access token and a refresh token that were generated via the SAML Authenticate API

pub fn saml_prepare_authentication<'b>(
&'a self
) -> SecuritySamlPrepareAuthentication<'a, 'b, ()>

Security Saml Prepare Authentication API

Creates a SAML authentication request

pub fn saml_service_provider_metadata<'b>(
&'a self,
parts: SecuritySamlServiceProviderMetadataParts<'b>
) -> SecuritySamlServiceProviderMetadata<'a, 'b>

Security Saml Service Provider Metadata API

Generates SAML metadata for the Elastic stack SAML 2.0 Service Provider