Struct elasticsearch::slm::Slm

pub struct Slm<'a> { /* private fields */ }

Namespace client for Snapshot Lifecycle Management APIs

Implementations

impl<'a> Slm<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Slm

pub fn transport(&self) -> &Transport

pub fn delete_lifecycle<'b>(
&'a self,
parts: SlmDeleteLifecycleParts<'b>
) -> SlmDeleteLifecycle<'a, 'b>

Slm Delete Lifecycle API

Deletes an existing snapshot lifecycle policy.

pub fn execute_lifecycle<'b>(
&'a self,
parts: SlmExecuteLifecycleParts<'b>
) -> SlmExecuteLifecycle<'a, 'b, ()>

Slm Execute Lifecycle API

Immediately creates a snapshot according to the lifecycle policy, without waiting for the scheduled time.

pub fn execute_retention<'b>(&'a self) -> SlmExecuteRetention<'a, 'b, ()>

Slm Execute Retention API

Deletes any snapshots that are expired according to the policy’s retention rules.

pub fn get_lifecycle<'b>(
&'a self,
parts: SlmGetLifecycleParts<'b>
) -> SlmGetLifecycle<'a, 'b>

Slm Get Lifecycle API

Retrieves one or more snapshot lifecycle policy definitions and information about the latest snapshot attempts.

pub fn get_stats<'b>(&'a self) -> SlmGetStats<'a, 'b>

Slm Get Stats API

Returns global and policy-level statistics about actions taken by snapshot lifecycle management.

pub fn get_status<'b>(&'a self) -> SlmGetStatus<'a, 'b>

Slm Get Status API

Retrieves the status of snapshot lifecycle management (SLM).

pub fn put_lifecycle<'b>(
&'a self,
parts: SlmPutLifecycleParts<'b>
) -> SlmPutLifecycle<'a, 'b, ()>

Slm Put Lifecycle API

Creates or updates a snapshot lifecycle policy.

pub fn start<'b>(&'a self) -> SlmStart<'a, 'b, ()>

Slm Start API

Turns on snapshot lifecycle management (SLM).

pub fn stop<'b>(&'a self) -> SlmStop<'a, 'b, ()>

Slm Stop API

Turns off snapshot lifecycle management (SLM).