Struct elasticsearch::Elasticsearch

pub struct Elasticsearch { /* fields omitted */ }

Root client for top level APIs

Implementations

impl Elasticsearch

pub fn async_search(&self) -> AsyncSearch<'_>

Creates a namespace client for AsyncSearch APIs

impl Elasticsearch

pub fn autoscaling(&self) -> Autoscaling<'_>

Creates a namespace client for Autoscaling APIs

impl Elasticsearch

pub fn cat(&self) -> Cat<'_>

Creates a namespace client for Cat APIs

impl Elasticsearch

pub fn ccr(&self) -> Ccr<'_>

Creates a namespace client for Cross Cluster Replication APIs

impl Elasticsearch

pub fn cluster(&self) -> Cluster<'_>

Creates a namespace client for Cluster APIs

impl Elasticsearch

pub fn dangling_indices(&self) -> DanglingIndices<'_>

Creates a namespace client for DanglingIndices APIs

impl Elasticsearch

pub fn enrich(&self) -> Enrich<'_>

Creates a namespace client for Enrich APIs

impl Elasticsearch

pub fn eql(&self) -> Eql<'_>

Creates a namespace client for Eql APIs

impl Elasticsearch

pub fn features(&self) -> Features<'_>

Creates a namespace client for Features APIs

impl Elasticsearch

pub fn graph(&self) -> Graph<'_>

Creates a namespace client for Graph APIs

impl Elasticsearch

pub fn ilm(&self) -> Ilm<'_>

Creates a namespace client for Index Lifecycle Management APIs

impl Elasticsearch

pub fn indices(&self) -> Indices<'_>

Creates a namespace client for Indices APIs

impl Elasticsearch

pub fn ingest(&self) -> Ingest<'_>

Creates a namespace client for Ingest APIs

impl Elasticsearch

pub fn license(&self) -> License<'_>

Creates a namespace client for License APIs

impl Elasticsearch

pub fn logstash(&self) -> Logstash<'_>

Creates a namespace client for Logstash APIs

impl Elasticsearch

pub fn migration(&self) -> Migration<'_>

Creates a namespace client for Migration APIs

impl Elasticsearch

pub fn ml(&self) -> Ml<'_>

Creates a namespace client for Machine Learning APIs

impl Elasticsearch

pub fn nodes(&self) -> Nodes<'_>

Creates a namespace client for Nodes APIs

impl Elasticsearch

pub fn security(&self) -> Security<'_>

Creates a namespace client for Security APIs

impl Elasticsearch

pub fn slm(&self) -> Slm<'_>

Creates a namespace client for Snapshot Lifecycle Management APIs

impl Elasticsearch

pub fn snapshot(&self) -> Snapshot<'_>

Creates a namespace client for Snapshot APIs

impl Elasticsearch

pub fn sql(&self) -> Sql<'_>

Creates a namespace client for Sql APIs

impl Elasticsearch

pub fn ssl(&self) -> Ssl<'_>

Creates a namespace client for Ssl APIs

impl Elasticsearch

pub fn text_structure(&self) -> TextStructure<'_>

Creates a namespace client for TextStructure APIs

impl Elasticsearch

pub fn transform(&self) -> Transform<'_>

Creates a namespace client for Transform APIs

impl Elasticsearch

pub fn watcher(&self) -> Watcher<'_>

Creates a namespace client for Watcher APIs

impl Elasticsearch

pub fn xpack(&self) -> Xpack<'_>

Creates a namespace client for X-Pack APIs

impl Elasticsearch

pub fn new(transport: Transport) -> Self

Creates a new instance of the root client

pub fn transport(&self) -> &Transport

Gets the transport of the client

pub async fn send<B, Q>(
&self,
method: Method,
path: &str,
headers: HeaderMap,
query_string: Option<&Q>,
body: Option<B>,
timeout: Option<Duration>
) -> Result<Response, Error> where
B: Body,
Q: Serialize + ?Sized,

Creates an asynchronous request that can be awaited

Accepts the HTTP method and relative path to an API, and optional query string and body.

impl Elasticsearch

pub fn bulk<'a, 'b>(&'a self, parts: BulkParts<'b>) -> Bulk<'a, 'b, ()>

Bulk API

Allows to perform multiple index/update/delete operations in a single request.

pub fn clear_scroll<'a, 'b>(
&'a self,
parts: ClearScrollParts<'b>
) -> ClearScroll<'a, 'b, ()>

Clear Scroll API

Explicitly clears the search context for a scroll.

pub fn close_point_in_time<'a, 'b>(&'a self) -> ClosePointInTime<'a, 'b, ()>

Close Point In Time API

Close a point in time

pub fn count<'a, 'b>(&'a self, parts: CountParts<'b>) -> Count<'a, 'b, ()>

Count API

Returns number of documents matching a query.

pub fn create<'a, 'b>(&'a self, parts: CreateParts<'b>) -> Create<'a, 'b, ()>

Create API

Creates a new document in the index.

Returns a 409 response when a document with a same ID already exists in the index.

pub fn delete<'a, 'b>(&'a self, parts: DeleteParts<'b>) -> Delete<'a, 'b>

Delete API

Removes a document from the index.

pub fn delete_by_query<'a, 'b>(
&'a self,
parts: DeleteByQueryParts<'b>
) -> DeleteByQuery<'a, 'b, ()>

Delete By Query API

Deletes documents matching the provided query.

pub fn delete_by_query_rethrottle<'a, 'b>(
&'a self,
parts: DeleteByQueryRethrottleParts<'b>
) -> DeleteByQueryRethrottle<'a, 'b, ()>

Delete By Query Rethrottle API

Changes the number of requests per second for a particular Delete By Query operation.

pub fn delete_script<'a, 'b>(
&'a self,
parts: DeleteScriptParts<'b>
) -> DeleteScript<'a, 'b>

Delete Script API

Deletes a script.

pub fn exists<'a, 'b>(&'a self, parts: ExistsParts<'b>) -> Exists<'a, 'b>

Exists API

Returns information about whether a document exists in an index.

pub fn exists_source<'a, 'b>(
&'a self,
parts: ExistsSourceParts<'b>
) -> ExistsSource<'a, 'b>

Exists Source API

Returns information about whether a document source exists in an index.

pub fn explain<'a, 'b>(&'a self, parts: ExplainParts<'b>) -> Explain<'a, 'b, ()>

Explain API

Returns information about why a specific matches (or doesn’t match) a query.

pub fn field_caps<'a, 'b>(
&'a self,
parts: FieldCapsParts<'b>
) -> FieldCaps<'a, 'b, ()>

Field Caps API

Returns the information about the capabilities of fields among multiple indices.

pub fn get<'a, 'b>(&'a self, parts: GetParts<'b>) -> Get<'a, 'b>

Get API

Returns a document.

pub fn get_script<'a, 'b>(
&'a self,
parts: GetScriptParts<'b>
) -> GetScript<'a, 'b>

Get Script API

Returns a script.

pub fn get_source<'a, 'b>(
&'a self,
parts: GetSourceParts<'b>
) -> GetSource<'a, 'b>

Get Source API

Returns the source of a document.

pub fn index<'a, 'b>(&'a self, parts: IndexParts<'b>) -> Index<'a, 'b, ()>

Index API

Creates or updates a document in an index.

pub fn info<'a, 'b>(&'a self) -> Info<'a, 'b>

Info API

Returns basic information about the cluster.

pub fn mget<'a, 'b>(&'a self, parts: MgetParts<'b>) -> Mget<'a, 'b, ()>

Mget API

Allows to get multiple documents in one request.

pub fn msearch<'a, 'b>(&'a self, parts: MsearchParts<'b>) -> Msearch<'a, 'b, ()>

Msearch API

Allows to execute several search operations in one request.

pub fn msearch_template<'a, 'b>(
&'a self,
parts: MsearchTemplateParts<'b>
) -> MsearchTemplate<'a, 'b, ()>

Msearch Template API

Allows to execute several search template operations in one request.

pub fn mtermvectors<'a, 'b>(
&'a self,
parts: MtermvectorsParts<'b>
) -> Mtermvectors<'a, 'b, ()>

Mtermvectors API

Returns multiple termvectors in one request.

pub fn open_point_in_time<'a, 'b>(
&'a self,
parts: OpenPointInTimeParts<'b>
) -> OpenPointInTime<'a, 'b, ()>

Open Point In Time API

Open a point in time that can be used in subsequent searches

pub fn ping<'a, 'b>(&'a self) -> Ping<'a, 'b>

Ping API

Returns whether the cluster is running.

pub fn put_script<'a, 'b>(
&'a self,
parts: PutScriptParts<'b>
) -> PutScript<'a, 'b, ()>

Put Script API

Creates or updates a script.

pub fn reindex<'a, 'b>(&'a self) -> Reindex<'a, 'b, ()>

Reindex API

Allows to copy documents from one index to another, optionally filtering the source documents by a query, changing the destination index settings, or fetching the documents from a remote cluster.

pub fn reindex_rethrottle<'a, 'b>(
&'a self,
parts: ReindexRethrottleParts<'b>
) -> ReindexRethrottle<'a, 'b, ()>

Reindex Rethrottle API

Changes the number of requests per second for a particular Reindex operation.

pub fn render_search_template<'a, 'b>(
&'a self,
parts: RenderSearchTemplateParts<'b>
) -> RenderSearchTemplate<'a, 'b, ()>

Render Search Template API

Allows to use the Mustache language to pre-render a search definition.

pub fn scroll<'a, 'b>(&'a self, parts: ScrollParts<'b>) -> Scroll<'a, 'b, ()>

Scroll API

Allows to retrieve a large numbers of results from a single search request.

Examples

To initiate a scroll, make search API call with a specified scroll timeout, then fetch the next set of hits using the _scroll_id returned in the response. Once no more hits are returned, clear the scroll.

let client = Elasticsearch::default();

fn print_hits(hits: &[Value]) {
    for hit in hits {
        println!(
            "id: '{}', source: '{}', score: '{}'",
            hit["_id"].as_str().unwrap(),
            hit["_source"],
            hit["_score"].as_f64().unwrap()
        );
    }
}

let scroll = "1m";
let mut response = client
    .search(SearchParts::Index(&["tweets"]))
    .scroll(scroll)
    .body(json!({
        "query": {
            "match": {
                "body": {
                    "query": "Elasticsearch rust",
                    "operator": "AND"
                }
            }
        }
    }))
    .send()
    .await?;

let mut response_body = response.json::<Value>().await?;
let mut scroll_id = response_body["_scroll_id"].as_str().unwrap();
let mut hits = response_body["hits"]["hits"].as_array().unwrap();

print_hits(hits);

while hits.len() > 0 {
    response = client
        .scroll(ScrollParts::None)
        .body(json!({
            "scroll": scroll,
            "scroll_id": scroll_id
        }))
        .send()
        .await?;

    response_body = response.json::<Value>().await?;
    scroll_id = response_body["_scroll_id"].as_str().unwrap();
    hits = response_body["hits"]["hits"].as_array().unwrap();
    print_hits(hits);
}

response = client
    .clear_scroll(ClearScrollParts::None)
    .body(json!({
        "scroll_id": scroll_id
    }))
    .send()
    .await?;
    

pub fn search<'a, 'b>(&'a self, parts: SearchParts<'b>) -> Search<'a, 'b, ()>

Search API

Returns results matching a query.

pub fn search_shards<'a, 'b>(
&'a self,
parts: SearchShardsParts<'b>
) -> SearchShards<'a, 'b, ()>

Search Shards API

Returns information about the indices and shards that a search request would be executed against.

pub fn search_template<'a, 'b>(
&'a self,
parts: SearchTemplateParts<'b>
) -> SearchTemplate<'a, 'b, ()>

Search Template API

Allows to use the Mustache language to pre-render a search definition.

pub fn termvectors<'a, 'b>(
&'a self,
parts: TermvectorsParts<'b>
) -> Termvectors<'a, 'b, ()>

Termvectors API

Returns information and statistics about terms in the fields of a particular document.

pub fn update<'a, 'b>(&'a self, parts: UpdateParts<'b>) -> Update<'a, 'b, ()>

Update API

Updates a document with a script or partial document.

pub fn update_by_query<'a, 'b>(
&'a self,
parts: UpdateByQueryParts<'b>
) -> UpdateByQuery<'a, 'b, ()>

Update By Query API

Performs an update on every document in the index without changing the source, for example to pick up a mapping change.

pub fn update_by_query_rethrottle<'a, 'b>(
&'a self,
parts: UpdateByQueryRethrottleParts<'b>
) -> UpdateByQueryRethrottle<'a, 'b, ()>

Update By Query Rethrottle API

Changes the number of requests per second for a particular Update By Query operation.