Struct elasticsearch::UpdateByQuery

pub struct UpdateByQuery<'a, 'b, B> { /* private fields */ }

Builder for the Update By Query API

Performs an update on every document in the index without changing the source, for example to pick up a mapping change.

Implementations

impl<'a, 'b, B> UpdateByQuery<'a, 'b, B> where
B: Body,

pub fn new(transport: &'a Transport, parts: UpdateByQueryParts<'b>) -> Self

Creates a new instance of UpdateByQuery with the specified API parts

pub fn _source(self, _source: &'b [&'b str]) -> Self

True or false to return the _source field or not, or a list of fields to return

pub fn _source_excludes(self, _source_excludes: &'b [&'b str]) -> Self

A list of fields to exclude from the returned _source field

pub fn _source_includes(self, _source_includes: &'b [&'b str]) -> Self

A list of fields to extract and return from the _source field

pub fn allow_no_indices(self, allow_no_indices: bool) -> Self

Whether to ignore if a wildcard indices expression resolves into no concrete indices. (This includes _all string or when no indices have been specified)

pub fn analyze_wildcard(self, analyze_wildcard: bool) -> Self

Specify whether wildcard and prefix queries should be analyzed (default: false)

pub fn analyzer(self, analyzer: &'b str) -> Self

The analyzer to use for the query string

pub fn body<T>(self, body: T) -> UpdateByQuery<'a, 'b, JsonBody<T>> where
T: Serialize,

The body for the API call

pub fn conflicts(self, conflicts: Conflicts) -> Self

What to do when the update by query hits version conflicts?

pub fn default_operator(self, default_operator: DefaultOperator) -> Self

The default operator for query string query (AND or OR)

pub fn df(self, df: &'b str) -> Self

The field to use as default where no field prefix is given in the query string

pub fn error_trace(self, error_trace: bool) -> Self

Include the stack trace of returned errors.

pub fn expand_wildcards(self, expand_wildcards: &'b [ExpandWildcards]) -> Self

Whether to expand wildcard expression to concrete indices that are open, closed or both.

pub fn filter_path(self, filter_path: &'b [&'b str]) -> Self

A comma-separated list of filters used to reduce the response.

pub fn from(self, from: i64) -> Self

Starting offset (default: 0)

pub fn header(self, key: HeaderName, value: HeaderValue) -> Self

Adds a HTTP header

pub fn human(self, human: bool) -> Self

Return human readable values for statistics.

pub fn ignore_unavailable(self, ignore_unavailable: bool) -> Self

Whether specified concrete indices should be ignored when unavailable (missing or closed)

pub fn lenient(self, lenient: bool) -> Self

Specify whether format-based query failures (such as providing text to a numeric field) should be ignored

pub fn max_docs(self, max_docs: i64) -> Self

Maximum number of documents to process (default: all documents)

pub fn pipeline(self, pipeline: &'b str) -> Self

Ingest pipeline to set on index requests made by this action. (default: none)

pub fn preference(self, preference: &'b str) -> Self

Specify the node or shard the operation should be performed on (default: random)

pub fn pretty(self, pretty: bool) -> Self

Pretty format the returned JSON response.

pub fn q(self, q: &'b str) -> Self

Query in the Lucene query string syntax

pub fn refresh(self, refresh: bool) -> Self

Should the affected indexes be refreshed?

pub fn request_cache(self, request_cache: bool) -> Self

Specify if request cache should be used for this request or not, defaults to index level setting

pub fn request_timeout(self, timeout: Duration) -> Self

Sets a request timeout for this API call.

The timeout is applied from when the request starts connecting until the response body has finished.

pub fn requests_per_second(self, requests_per_second: i64) -> Self

The throttle to set on this request in sub-requests per second. -1 means no throttle.

pub fn routing(self, routing: &'b [&'b str]) -> Self

A comma-separated list of specific routing values

pub fn scroll(self, scroll: &'b str) -> Self

Specify how long a consistent view of the index should be maintained for scrolled search

pub fn scroll_size(self, scroll_size: i64) -> Self

Size on the scroll request powering the update by query

pub fn search_timeout(self, search_timeout: &'b str) -> Self

Explicit timeout for each search request. Defaults to no timeout.

pub fn search_type(self, search_type: SearchType) -> Self

Search operation type

pub fn size(self, size: i64) -> Self

Deprecated, please use max_docs instead

pub fn slices(self, slices: Slices) -> Self

The number of slices this task should be divided into. Defaults to 1, meaning the task isn’t sliced into subtasks. Can be set to auto.

pub fn sort(self, sort: &'b [&'b str]) -> Self

A comma-separated list of : pairs

pub fn source(self, source: &'b str) -> Self

The URL-encoded request definition. Useful for libraries that do not accept a request body for non-POST requests.

pub fn stats(self, stats: &'b [&'b str]) -> Self

Specific ‘tag’ of the request for logging and statistical purposes

pub fn terminate_after(self, terminate_after: i64) -> Self

The maximum number of documents to collect for each shard, upon reaching which the query execution will terminate early.

pub fn timeout(self, timeout: &'b str) -> Self

Time each individual bulk request should wait for shards that are unavailable.

pub fn version(self, version: bool) -> Self

Specify whether to return document version as part of a hit

pub fn version_type(self, version_type: bool) -> Self

Should the document increment the version number (internal) on hit or not (reindex)

pub fn wait_for_active_shards(self, wait_for_active_shards: &'b str) -> Self

Sets the number of shard copies that must be active before proceeding with the update by query operation. Defaults to 1, meaning the primary shard only. Set to all for all shard copies, otherwise set to any non-negative value less than or equal to the total number of copies for the shard (number of replicas + 1)

pub fn wait_for_completion(self, wait_for_completion: bool) -> Self

Should the request should block until the update by query operation is complete.

pub async fn send(self) -> Result<Response, Error>

Creates an asynchronous call to the Update By Query API that can be awaited