Module elasticsearch::watcher

Watcher (Alerting) APIs

Enable watching for changes or anomalies in data and perform the necessary actions in response, by creating and managing watches that take action based on a met condition.

Structs

Watcher

Namespace client for Watcher APIs

WatcherAckWatch

Builder for the Watcher Ack Watch API

WatcherActivateWatch

Builder for the Watcher Activate Watch API

WatcherDeactivateWatch

Builder for the Watcher Deactivate Watch API

WatcherDeleteWatch

Builder for the Watcher Delete Watch API

WatcherExecuteWatch

Builder for the Watcher Execute Watch API

WatcherGetWatch

Builder for the Watcher Get Watch API

WatcherPutWatch

Builder for the Watcher Put Watch API

WatcherQueryWatches

Builder for the Watcher Query Watches API

WatcherStart

Builder for the Watcher Start API

WatcherStats

Builder for the Watcher Stats API

WatcherStop

Builder for the Watcher Stop API

Enums

WatcherAckWatchParts

API parts for the Watcher Ack Watch API

WatcherActivateWatchParts

API parts for the Watcher Activate Watch API

WatcherDeactivateWatchParts

API parts for the Watcher Deactivate Watch API

WatcherDeleteWatchParts

API parts for the Watcher Delete Watch API

WatcherExecuteWatchParts

API parts for the Watcher Execute Watch API

WatcherGetWatchParts

API parts for the Watcher Get Watch API

WatcherPutWatchParts

API parts for the Watcher Put Watch API

WatcherQueryWatchesParts

API parts for the Watcher Query Watches API

WatcherStartParts

API parts for the Watcher Start API

WatcherStatsParts

API parts for the Watcher Stats API

WatcherStopParts

API parts for the Watcher Stop API