Struct elasticsearch::watcher::Watcher

pub struct Watcher<'a> { /* fields omitted */ }

Namespace client for Watcher APIs

Implementations

impl<'a> Watcher<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Watcher

pub fn transport(&self) -> &Transport

pub fn ack_watch<'b>(
&'a self,
parts: WatcherAckWatchParts<'b>
) -> WatcherAckWatch<'a, 'b, ()>

Watcher Ack Watch API

Acknowledges a watch, manually throttling the execution of the watch’s actions.

pub fn activate_watch<'b>(
&'a self,
parts: WatcherActivateWatchParts<'b>
) -> WatcherActivateWatch<'a, 'b, ()>

Watcher Activate Watch API

Activates a currently inactive watch.

pub fn deactivate_watch<'b>(
&'a self,
parts: WatcherDeactivateWatchParts<'b>
) -> WatcherDeactivateWatch<'a, 'b, ()>

Watcher Deactivate Watch API

Deactivates a currently active watch.

pub fn delete_watch<'b>(
&'a self,
parts: WatcherDeleteWatchParts<'b>
) -> WatcherDeleteWatch<'a, 'b>

Watcher Delete Watch API

Removes a watch from Watcher.

pub fn execute_watch<'b>(
&'a self,
parts: WatcherExecuteWatchParts<'b>
) -> WatcherExecuteWatch<'a, 'b, ()>

Watcher Execute Watch API

Forces the execution of a stored watch.

pub fn get_watch<'b>(
&'a self,
parts: WatcherGetWatchParts<'b>
) -> WatcherGetWatch<'a, 'b>

Watcher Get Watch API

Retrieves a watch by its ID.

pub fn put_watch<'b>(
&'a self,
parts: WatcherPutWatchParts<'b>
) -> WatcherPutWatch<'a, 'b, ()>

Watcher Put Watch API

Creates a new watch, or updates an existing one.

pub fn query_watches<'b>(&'a self) -> WatcherQueryWatches<'a, 'b, ()>

Watcher Query Watches API

Retrieves stored watches.

pub fn start<'b>(&'a self) -> WatcherStart<'a, 'b, ()>

Watcher Start API

Starts Watcher if it is not already running.

pub fn stats<'b>(&'a self, parts: WatcherStatsParts<'b>) -> WatcherStats<'a, 'b>

Watcher Stats API

Retrieves the current Watcher metrics.

pub fn stop<'b>(&'a self) -> WatcherStop<'a, 'b, ()>

Watcher Stop API

Stops Watcher if it is running.