Struct elasticsearch::cat::Cat

pub struct Cat<'a> { /* private fields */ }

Namespace client for Cat APIs

Implementations

impl<'a> Cat<'a>

pub fn new(transport: &'a Transport) -> Self

Creates a new instance of Cat

pub fn transport(&self) -> &Transport

pub fn aliases<'b>(&'a self, parts: CatAliasesParts<'b>) -> CatAliases<'a, 'b>

Cat Aliases API

Shows information about currently configured aliases to indices including filter and routing infos.

pub fn allocation<'b>(
&'a self,
parts: CatAllocationParts<'b>
) -> CatAllocation<'a, 'b>

Cat Allocation API

Provides a snapshot of how many shards are allocated to each data node and how much disk space they are using.

pub fn count<'b>(&'a self, parts: CatCountParts<'b>) -> CatCount<'a, 'b>

Cat Count API

Provides quick access to the document count of the entire cluster, or individual indices.

pub fn fielddata<'b>(
&'a self,
parts: CatFielddataParts<'b>
) -> CatFielddata<'a, 'b>

Cat Fielddata API

Shows how much heap memory is currently being used by fielddata on every data node in the cluster.

pub fn health<'b>(&'a self) -> CatHealth<'a, 'b>

Cat Health API

Returns a concise representation of the cluster health.

pub fn help<'b>(&'a self) -> CatHelp<'a, 'b>

Cat Help API

Returns help for the Cat APIs.

pub fn indices<'b>(&'a self, parts: CatIndicesParts<'b>) -> CatIndices<'a, 'b>

Cat Indices API

Returns information about indices: number of primaries and replicas, document counts, disk size, …

pub fn master<'b>(&'a self) -> CatMaster<'a, 'b>

Cat Master API

Returns information about the master node.

pub fn ml_data_frame_analytics<'b>(
&'a self,
parts: CatMlDataFrameAnalyticsParts<'b>
) -> CatMlDataFrameAnalytics<'a, 'b>

Cat Ml Data Frame Analytics API

Gets configuration and usage information about data frame analytics jobs.

pub fn ml_datafeeds<'b>(
&'a self,
parts: CatMlDatafeedsParts<'b>
) -> CatMlDatafeeds<'a, 'b>

Cat Ml Datafeeds API

Gets configuration and usage information about datafeeds.

pub fn ml_jobs<'b>(&'a self, parts: CatMlJobsParts<'b>) -> CatMlJobs<'a, 'b>

Cat Ml Jobs API

Gets configuration and usage information about anomaly detection jobs.

pub fn ml_trained_models<'b>(
&'a self,
parts: CatMlTrainedModelsParts<'b>
) -> CatMlTrainedModels<'a, 'b>

Cat Ml Trained Models API

Gets configuration and usage information about inference trained models.

pub fn nodeattrs<'b>(&'a self) -> CatNodeattrs<'a, 'b>

Cat Nodeattrs API

Returns information about custom node attributes.

pub fn nodes<'b>(&'a self) -> CatNodes<'a, 'b>

Cat Nodes API

Returns basic statistics about performance of cluster nodes.

pub fn pending_tasks<'b>(&'a self) -> CatPendingTasks<'a, 'b>

Cat Pending Tasks API

Returns a concise representation of the cluster pending tasks.

pub fn plugins<'b>(&'a self) -> CatPlugins<'a, 'b>

Cat Plugins API

Returns information about installed plugins across nodes node.

pub fn recovery<'b>(
&'a self,
parts: CatRecoveryParts<'b>
) -> CatRecovery<'a, 'b>

Cat Recovery API

Returns information about index shard recoveries, both on-going completed.

pub fn repositories<'b>(&'a self) -> CatRepositories<'a, 'b>

Cat Repositories API

Returns information about snapshot repositories registered in the cluster.

pub fn segments<'b>(
&'a self,
parts: CatSegmentsParts<'b>
) -> CatSegments<'a, 'b>

Cat Segments API

Provides low-level information about the segments in the shards of an index.

pub fn shards<'b>(&'a self, parts: CatShardsParts<'b>) -> CatShards<'a, 'b>

Cat Shards API

Provides a detailed view of shard allocation on nodes.

pub fn snapshots<'b>(
&'a self,
parts: CatSnapshotsParts<'b>
) -> CatSnapshots<'a, 'b>

Cat Snapshots API

Returns all snapshots in a specific repository.

pub fn tasks<'b>(&'a self) -> CatTasks<'a, 'b>

Cat Tasks API

Returns information about the tasks currently executing on one or more nodes in the cluster.

pub fn templates<'b>(
&'a self,
parts: CatTemplatesParts<'b>
) -> CatTemplates<'a, 'b>

Cat Templates API

Returns information about existing templates.

pub fn thread_pool<'b>(
&'a self,
parts: CatThreadPoolParts<'b>
) -> CatThreadPool<'a, 'b>

Cat Thread Pool API

Returns cluster-wide thread pool statistics per node. By default the active, queue and rejected statistics are returned for all thread pools.

pub fn transforms<'b>(
&'a self,
parts: CatTransformsParts<'b>
) -> CatTransforms<'a, 'b>

Cat Transforms API

Gets configuration and usage information about transforms.