Module elasticsearch::dangling_indices

Dangling Index APIs

If Elasticsearch encounters index data that is absent from the current cluster state, those indices are considered to be dangling. For example, this can happen if you delete more than cluster.indices.tombstones.size number of indices while an Elasticsearch node is offline.

The dangling indices APIs can list, import and delete dangling indices.

Structs

DanglingIndices

Namespace client for DanglingIndices APIs

DanglingIndicesDeleteDanglingIndex

Builder for the Dangling Indices Delete Dangling Index API

DanglingIndicesImportDanglingIndex

Builder for the Dangling Indices Import Dangling Index API

DanglingIndicesListDanglingIndices

Builder for the Dangling Indices List Dangling Indices API

Enums

DanglingIndicesDeleteDanglingIndexParts

API parts for the Dangling Indices Delete Dangling Index API

DanglingIndicesImportDanglingIndexParts

API parts for the Dangling Indices Import Dangling Index API

DanglingIndicesListDanglingIndicesParts

API parts for the Dangling Indices List Dangling Indices API