Module elasticsearch::license

Licensing APIs

Manage Elastic Stack license, including

Retrieve, update or delete a license

Start a 30 day trial of the Platinum license features

Start indefinite use of the Basic license features

Get the status of trial and basic license features

Structs

License

Namespace client for License APIs

LicenseDelete

Builder for the License Delete API

LicenseGet

Builder for the License Get API

LicenseGetBasicStatus

Builder for the License Get Basic Status API

LicenseGetTrialStatus

Builder for the License Get Trial Status API

LicensePost

Builder for the License Post API

LicensePostStartBasic

Builder for the License Post Start Basic API

LicensePostStartTrial

Builder for the License Post Start Trial API

Enums

LicenseDeleteParts

API parts for the License Delete API

LicenseGetBasicStatusParts

API parts for the License Get Basic Status API

LicenseGetParts

API parts for the License Get API

LicenseGetTrialStatusParts

API parts for the License Get Trial Status API

LicensePostParts

API parts for the License Post API

LicensePostStartBasicParts

API parts for the License Post Start Basic API

LicensePostStartTrialParts

API parts for the License Post Start Trial API