Module elasticsearch::slm

Snapshot Lifecycle Management APIs

Manage policies for the time and frequency of automatic snapshots. Snapshot Lifecycle Management (SLM) is related to Index Lifecycle Management, however, instead of managing a lifecycle of actions that are performed on a single index, SLM allows configuring policies spanning multiple indices. Snapshot Lifecycle Management can also perform deletion of older snapshots based on a configurable retention policy.

Structs

Slm

Namespace client for Snapshot Lifecycle Management APIs

SlmDeleteLifecycle

Builder for the Slm Delete Lifecycle API

SlmExecuteLifecycle

Builder for the Slm Execute Lifecycle API

SlmExecuteRetention

Builder for the Slm Execute Retention API

SlmGetLifecycle

Builder for the Slm Get Lifecycle API

SlmGetStats

Builder for the Slm Get Stats API

SlmGetStatus

Builder for the Slm Get Status API

SlmPutLifecycle

Builder for the Slm Put Lifecycle API

SlmStart

Builder for the Slm Start API

SlmStop

Builder for the Slm Stop API

Enums

SlmDeleteLifecycleParts

API parts for the Slm Delete Lifecycle API

SlmExecuteLifecycleParts

API parts for the Slm Execute Lifecycle API

SlmExecuteRetentionParts

API parts for the Slm Execute Retention API

SlmGetLifecycleParts

API parts for the Slm Get Lifecycle API

SlmGetStatsParts

API parts for the Slm Get Stats API

SlmGetStatusParts

API parts for the Slm Get Status API

SlmPutLifecycleParts

API parts for the Slm Put Lifecycle API

SlmStartParts

API parts for the Slm Start API

SlmStopParts

API parts for the Slm Stop API