Module elasticsearch::snapshot

Snapshot APIs

Manage snapshots taken from a running Elasticsearch cluster. A snapshot is a backup of individual indices or the entire cluster, stored in a repository on a shared filesystem or a remote repository on S3, HDFS, Azure, Google Cloud storage, and more.

Structs

Snapshot

Namespace client for Snapshot APIs

SnapshotCleanupRepository

Builder for the Snapshot Cleanup Repository API

SnapshotClone

Builder for the Snapshot Clone API

SnapshotCreate

Builder for the Snapshot Create API

SnapshotCreateRepository

Builder for the Snapshot Create Repository API

SnapshotDelete

Builder for the Snapshot Delete API

SnapshotDeleteRepository

Builder for the Snapshot Delete Repository API

SnapshotGet

Builder for the Snapshot Get API

SnapshotGetRepository

Builder for the Snapshot Get Repository API

SnapshotRepositoryAnalyze

Builder for the Snapshot Repository Analyze API

SnapshotRestore

Builder for the Snapshot Restore API

SnapshotStatus

Builder for the Snapshot Status API

SnapshotVerifyRepository

Builder for the Snapshot Verify Repository API

Enums

SnapshotCleanupRepositoryParts

API parts for the Snapshot Cleanup Repository API

SnapshotCloneParts

API parts for the Snapshot Clone API

SnapshotCreateParts

API parts for the Snapshot Create API

SnapshotCreateRepositoryParts

API parts for the Snapshot Create Repository API

SnapshotDeleteParts

API parts for the Snapshot Delete API

SnapshotDeleteRepositoryParts

API parts for the Snapshot Delete Repository API

SnapshotGetParts

API parts for the Snapshot Get API

SnapshotGetRepositoryParts

API parts for the Snapshot Get Repository API

SnapshotRepositoryAnalyzeParts

API parts for the Snapshot Repository Analyze API

SnapshotRestoreParts

API parts for the Snapshot Restore API

SnapshotStatusParts

API parts for the Snapshot Status API

SnapshotVerifyRepositoryParts

API parts for the Snapshot Verify Repository API