Module elasticsearch::sql

SQL APIs

Execute SQL queries against Elasticsearch indices and return results in tabular format.

Structs

Sql

Namespace client for Sql APIs

SqlClearCursor

Builder for the Sql Clear Cursor API

SqlDeleteAsync

Builder for the Sql Delete Async API

SqlGetAsync

Builder for the Sql Get Async API

SqlGetAsyncStatus

Builder for the Sql Get Async Status API

SqlQuery

Builder for the Sql Query API

SqlTranslate

Builder for the Sql Translate API

Enums

SqlClearCursorParts

API parts for the Sql Clear Cursor API

SqlDeleteAsyncParts

API parts for the Sql Delete Async API

SqlGetAsyncParts

API parts for the Sql Get Async API

SqlGetAsyncStatusParts

API parts for the Sql Get Async Status API

SqlQueryParts

API parts for the Sql Query API

SqlTranslateParts

API parts for the Sql Translate API