Module elasticsearch::text_structure

Text structure APIs

Determines the structure of text and other information that will be useful to import its contents to an Elasticsearch index.

Structs

TextStructure

Namespace client for TextStructure APIs

TextStructureFindStructure

Builder for the Text Structure Find Structure API

Enums

TextStructureFindStructureParts

API parts for the Text Structure Find Structure API