Module elasticsearch::transform

Transform APIs

Transforms can be used to copy data from source indices, transforms it, and persists it into an entity-centric destination index.

Structs

Transform

Namespace client for Transform APIs

TransformDeleteTransform

Builder for the Transform Delete Transform API

TransformGetTransform

Builder for the Transform Get Transform API

TransformGetTransformStats

Builder for the Transform Get Transform Stats API

TransformPreviewTransform

Builder for the Transform Preview Transform API

TransformPutTransform

Builder for the Transform Put Transform API

TransformStartTransform

Builder for the Transform Start Transform API

TransformStopTransform

Builder for the Transform Stop Transform API

TransformUpdateTransform

Builder for the Transform Update Transform API

Enums

TransformDeleteTransformParts

API parts for the Transform Delete Transform API

TransformGetTransformParts

API parts for the Transform Get Transform API

TransformGetTransformStatsParts

API parts for the Transform Get Transform Stats API

TransformPreviewTransformParts

API parts for the Transform Preview Transform API

TransformPutTransformParts

API parts for the Transform Put Transform API

TransformStartTransformParts

API parts for the Transform Start Transform API

TransformStopTransformParts

API parts for the Transform Stop Transform API

TransformUpdateTransformParts

API parts for the Transform Update Transform API